% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{add.remove.MK.cols}
\alias{add.remove.MK.cols}
\title{Harmonize columns for enterprise data.}
\usage{
add.remove.MK.cols(
  data,
  tab,
  cols,
  rows,
  order = c("col", "row"),
  sep = "_",
  verbose = FALSE
)
}
\description{
Harmonize columns for enterprise data.
}
\details{
This function creates a vector of 'Merkmale' that contain all combinations of Merkmals-Spalten and Merkmals-Zeilen for a given Merkmals-Tabelle.
Example: M100.0100.10000
It removes all columns in dat that are not given in the list.
It adds all columns in dat if they are not already there (and sets their value equal 0).
In the end, it sorts the colums of dat by order(cols,rows) or order(rows, cols).
If comment==TRUE, it is reported, which columns are removed and added to dat.
}
