% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{addColsBefore}
\alias{addColsBefore}
\title{Add column(s) to a data.frame/matrix/list before a certain other column (or list place).}
\usage{
addColsBefore(beforeName, addNames = NULL, values = NULL, data)
}
\arguments{
\item{beforeName}{Before this column name in \code{data}, add the other columns given in \code{addNames}.}

\item{addNames}{Optional: The names to be added/rearranged. This only has to be specified when \code{values} has no names.}

\item{values}{Optional: The values to be added into \code{data}. This only has to be specified when new data should be added rather than only rearranging columns.}

\item{data}{The data.frame, list, or matrix in which the rearrangement should take place.}
}
\value{
\code{data} with rearranged/added columns.
}
\description{
Add column(s) to a data.frame/matrix/list before a certain other column (or list place).
}
\examples{
dat <- data.frame(a = 1:3, b = 2:4, z = 3:5)
beforeName <- "a"
values <- list(c = 5:7)

# The following examples work without errors.
addColsBefore(beforeName = beforeName, addNames = NULL, values = values, data = dat)
addColsBefore(beforeName = beforeName, addNames = NULL, values = as.data.frame(values), data = dat)
addColsBefore(beforeName = beforeName, addNames = NULL, values = as.matrix(as.data.frame(values)), data = dat)
addColsBefore(beforeName = beforeName, addNames = "x", values = values, data = dat)
addColsBefore(beforeName = "z", addNames = "x", values = values, data = dat)
addColsBefore(beforeName = "z", addNames = "z", data = dat)
# The following will only rearrange.
addColsBefore(beforeName = beforeName, addNames = "z", values = NULL, data = dat)
# The following will rearrange and give new values to column "z".
addColsBefore(beforeName = beforeName, addNames = "z", values = 1:3, data = dat)

# The following examples will yield errors.
addColsBefore(beforeName = beforeName, addNames = NULL, values = 1:3, data = dat)
values <- list(c = 5:7, d = 6:7)
addColsBefore(beforeName = beforeName, addNames = NULL, values = values, data = dat)

# In lists, it is allowed that every list place has different length. Hence, no error.
dat <- list(a = 1:3, b = 2:4, z = 1)
addColsBefore(beforeName = beforeName, addNames = NULL, values = values, data = dat)
}
