% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{agsPath}
\alias{agsPath}
\title{Convert paths in Agroscope domain from Windows to RStudio format.}
\usage{
agsPath(path)
}
\arguments{
\item{path}{The path to convert.}
}
\value{
The converted path.
}
\description{
Convert paths in Agroscope domain from Windows to RStudio format.
}
\seealso{
\code{\link[agsGitlabUtils:agsPath]{agsGitlabUtils::agsPath}}
}
\author{
Daniel Hoop
}
