% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{ahv.anteil.betrieb}
\alias{ahv.anteil.betrieb}
\title{Get the AHV Anteil that is payed by the farm (not the farmer).}
\usage{
ahv.anteil.betrieb(jahr = NULL)
}
\description{
E.g. 0.60 means that the farm pays 60\% of the BVG, and the farmer pays 40\% him/herself.
}
\author{
Daniel Hoop
}
