% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{ahv.satz.BG}
\alias{ahv.satz.BG}
\title{Get the AHV Satz that is assumed for Betriebsgemeinschaften.}
\usage{
ahv.satz.BG(jahr = NULL)
}
\description{
E.g. 0.125 means that the rate is 12.5\% of the farm income, i.e. \code{AHV_total = 0.125 * Landw_Einkommen}.
}
\author{
Daniel Hoop
}
