% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{assertDatasetEquality}
\alias{assertDatasetEquality}
\title{Asserts that two data.frames/matrices \code{dat1} and \code{dat2} are equal.}
\usage{
assertDatasetEquality(
  dat1,
  dat2,
  idCols = NULL,
  compareAvailObsOnly = FALSE,
  colsToCheck = NULL,
  regexToCheck = NULL,
  colsShowInReport = NULL,
  showAllDiffCols = FALSE,
  showBothValuesNotDiff = TRUE,
  tolerance = 1e-06,
  sepForDiff = " <|> ",
  outFile = NULL,
  openOutFile = FALSE,
  sep = ";",
  header = TRUE,
  ...
)
}
\arguments{
\item{dat1}{A data.frame/matrix or the path to a text file that can be read with 'read.table' or the path to a RData file that contains a single R object.}

\item{dat2}{A data.frame/matrix or the path to a text file that can be read with 'read.table' or the path to a RData file that contains a single R object.}

\item{idCols}{Optional: The columns that uniquely identify a observation in each row. If given, then only rows will compared which are available in both datasets. Otherwise, a different number of rows in \code{dat1} and \code{dat2} will be reported as difference.}

\item{compareAvailObsOnly}{If \code{idCols} is given, should only those observations be compared which are available in both datasets? Logical value.}

\item{colsToCheck}{The columns that should be checked for differences in dat1 and dat2.}

\item{regexToCheck}{A regular expression defining the columns that should be checked for differences in dat1 and dat2.}

\item{showBothValuesNotDiff}{Logical value indicating if (in case of disceprancy) both values should be shown instead of showig the difference.}

\item{tolerance}{The tolerance with which equality of numeric values should be checked.}

\item{sepForDiff}{The file to separate two different values. On the left side of the \code{sepForDiff} sign, the value from \code{dat1} is shown. On the right side, the value from \code{dat2} is shown.}

\item{outFile}{Optional: The name of the file that should be created which will contain all errorneous observations and columns.}

\item{sep}{The column delimiter used in read.table and write.table}

\item{header}{Logical value indicating if the files to read in have a header}

\item{...}{Additional arguments passed to \code{\link[utils:write.table]{utils::write.table}}}

\item{colsToShowInReport}{Additional columns that should be shown in the report (apart from the columns named in idCols and the errorneous cols).}
}
\value{
\code{NULL} if no differences were detected.
Else, a \code{data.frame} containing all errorneous rows and columns, if differences were detected. The \code{data.frame} contains only the differences in the cells.
}
\description{
If not, a warning is shown and the differences in the data sets is returned.
}
\details{
Either \code{colsToCheck} or \code{regexToCheck} must be given, otherwise an error is thrown.
}
\seealso{
\code{\link{assertAggregationConsistency}}
}
\author{
Daniel Hoop
}
