% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{availableCores}
\alias{availableCores}
\title{Calculate the number of available cores such that memory will not be exhausted by parallel processing.}
\usage{
availableCores(cpuBuffer = 0.2, memBuffer = 0.3, roundFunc = base::floor)
}
\arguments{
\item{cpuBuffer}{Numeric value specifying the share totally available processors that should not be used but left as buffer for other processes.}

\item{memBuffer}{Numeric value specifying the share of the total memory that should not be used but left as buffer for other processes.}

\item{roundFunc}{The rounding function for the number of cores. Default is \code{\link[base:floor]{base::floor}}}
}
\value{
The number of cores.
}
\description{
Calculate the number of available cores such that memory will not be exhausted by parallel processing.
}
