% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{balanced.panel.filtering}
\alias{balanced.panel.filtering}
\title{Filters a balanced panel data set such that the criterium is fulfilled in every year for each observation.}
\usage{
balanced.panel.filtering(
  criterium,
  data,
  id.vec,
  year.vec,
  output = c("logical", "ID"),
  NAasFALSE = TRUE
)
}
\arguments{
\item{criterium}{Expression, that formulates which filtering-criterium must be fulfilled. Colnames of the data must be used!}

\item{data}{Matrix or data.frame that contains all colnames which are used in the criterium expression.}

\item{id.vec}{Vector of IDs (preferrably something like data[,"ID"] )}

\item{year.vec}{Vector of years (preferrably something like data[,"year"] )}

\item{output}{Should the output be in form of a logical vector or IDs? --> Better choose "logical" in order not to have problems!}
}
\description{
Filters a balanced panel data set such that the criterium is fulfilled in every year for each observation.
}
\examples{
data <- as.data.frame(matrix(1:100, ncol=2))
colnames(data) <- c("a","b")
id.vec <- rep(1:25, 2)
year.vec <- rep.1b1(c(2003, 2004), 25)
cbind(data, sum=rowSums(data), id=id.vec, year=year.vec, filter=balanced.panel.filtering(a+b<110, data, id.vec, year.vec))
}
\author{
Daniel Hoop
}
