% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{balanced.panel}
\alias{balanced.panel}
\title{Creates a vector to filter observations to a balanced panel.}
\usage{
balanced.panel(
  id,
  year,
  YEAR = sort(unique(year)),
  index = NULL,
  nYEARmin = length(unique(YEAR)),
  output = c("logical", "ID")
)
}
\arguments{
\item{id}{Vector of IDs}

\item{year}{Vector of year (same length as ID)}

\item{YEAR}{Years that should be selected}

\item{index}{Index for which the balancing should be done.}

\item{output}{Logical vector or IDs?}

\item{nYearmin}{Minimum number of years to be selected for the pseudo balanced panel}
}
\description{
Creates a vector to filter observations to a balanced panel.
}
\examples{
id <- sample(1:10,100,TRUE)
year <- sample(2000:2002,100,TRUE)
YEAR <- c(2000,2001);
balanced.panel(id = id, year = year, YEAR = YEAR, output = "logical")
}
\author{
Daniel Hoop
}
