% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{by.add.df.cols}
\alias{by.add.df.cols}
\title{Uses \code{by()} over \code{data[,relevantColnames]} with \code{INDICES} and a defined function \code{FUN}.}
\usage{
\method{by}{add.df.cols}(
  data,
  relevantColnames = NULL,
  INDICES,
  FUN,
  showWarnings = TRUE,
  autoColSubset = FALSE
)
}
\arguments{
\item{data}{The data frame over which \code{\link[base:by]{base::by}} should be applied.}

\item{relevantColnames}{The relevant colnames that are needed for the calculations in FUN.}

\item{INDICES}{The indices for the application of \code{\link[base:by]{base::by}}. This will be pasted to a signle string if \code{!is.null(dim(INDICES))}.}

\item{FUN}{The function to apply within \code{\link[base:by]{base::by}}.}

\item{autoColSubset}{Boolean value indicating if only the relevant colnames from data should be used in the calculations in order to speed up the calculations. *Experimental*}
}
\description{
It is designed to be faster than \code{\link[base:by]{base::by}} over the whole \code{data.frame} because is takes only these columns into \code{\link[base:by]{base::by}} that are really needed for the function.
After the calculation a data.frame is returned instead of the usual \code{list} that is returned by the \code{\link[base:by]{base::by}} function.
The initial order of the rows in \code{data} is kept in the result. The additionally calculated columns from \code{FUN} are added to the original data.frame that was given in \code{data}.
}
\details{
The same can be accomplished with much less code, using the package \code{dplyr}.
}
\examples{
by(data.frame(a = 1:10, b = 1:10),
   INDICES=c(1,1,1,1,1,2,2,2,2,2),
   function(x) {
     x[,"count"] <- nrow(x)
     x[,"sum"] <- sum(c(x[,"a"], x[,"b"]))
     return (x)
   })
}
\seealso{
\code{\link[base:by]{base::by}}
}
\author{
Daniel Hoop
}
\keyword{internal}
