% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{c.dimnames}
\alias{c.dimnames}
\title{Integrates the dimnames of an array into the array itself}
\usage{
\method{c}{dimnames}(array, sep.sign = "_")
}
\arguments{
\item{array}{The array of which the dimnames should be integrated.}

\item{sep.sign}{The separator between the integrated dimnames.}
}
\description{
Integrates the dimnames of an array into the array itself
}
\examples{
# Example 1 - matrix with row & colnames
array <- matrix(0, ncol=3, nrow=3); dimnames(array) <- list(1:3, c("a","b","c"));
c.dimnames(array)

# Example 2 - matrix without rownames
array <- matrix(0, ncol=3, nrow=3); colnames(array) <- c("a","b","c");
c.dimnames(array)
# Example 3 - 3 dimensional array
array <- array(0, dim=c(3,4,5), dimnames=list(c("c","b","a"),c(4,3,2,1),c("z","y","x","v","u")))
c.dimnames(array)
}
\author{
Daniel Hoop
}
