% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{c.matrices}
\alias{c.matrices}
\title{Combine multiple matrices into one large matrix.}
\usage{
\method{c}{matrices}(
  ...,
  fill = "",
  nbreak = 0,
  aligned = c("left", "right"),
  integrate.dimnames = c("no", "rowcol", "row", "col"),
  func = c("rbind", "cbind")
)
}
\arguments{
\item{...}{The matrices to be combined.}

\item{fill}{The character that will be written into the separating rows/cols that were added between the combined matrices.}

\item{nbreak}{The number of rows/cols that should be added between the combined matrices.}

\item{aligned}{If the matrices don't have the same number of columns/rows, then the rest of the empty cells will be filled with \code{""}. Should all matrices be left- or right-alinged?}

\item{integrate.dimnames}{Character specifying, if the dimnaes of the matrices should be integrated into the resulting matrix. This can also be done manually by using the function \code{\link{dimnames.to.mat}}.}

\item{func}{The function to be applied. \code{"rbind"} or \code{"cbind"}.}
}
\description{
Combine multiple matrices into one large matrix.
}
\examples{
x <- data.frame(testCol=1:10, testC=10:1)
y <- data.frame(testCol2=1:5, testCol3=2:6, testcol4=11:15)
z <- 1:3; func=c("rbind","cbind")[1]; fill=""
c.matrices(x, y, z, nbreak = 1, aligned = "left", integrate.dimnames = "rowcol")
c.matrices(list(x, y, z), nbreak = 1, aligned = "right", integrate.dimnames = "col")
}
\author{
Daniel Hoop
}
