% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{calc.I.cols}
\alias{calc.I.cols}
\title{Calculates the value of all columns with colnames looking like I(a+b) in a matrix/data.frame/list.}
\usage{
calc.I.cols(
  data,
  edit.I.colnames = FALSE,
  del.I.help.columns = FALSE,
  I.help.columns = NULL
)
}
\arguments{
\item{data}{\code{matrix}/\code{data.frame}/\code{list}. Columns to be calculated must have names like \code{"I(a+b)"}, \code{"I(a/b)"}. However, \code{"a+b"} will not work.}

\item{edit.I.colnames}{Should the brackets I() in the colnames be removed after the calculations? E.g. \code{"I(a+b)"} will later be called \code{"a+b"}.}

\item{I.help.columns}{If del.I.help.colums=TRUE: The list of the columns to be deleted can be specified. Otherwise the help columns are dectected automatically.}

\item{del.I.help.colums}{If e.g. \code{"I(a+b)"} is calculated. Should the columns \code{"a"} and \code{"b"} be removed after the calculation because they are of no interest?}
}
\description{
Calculates the value of all columns with colnames looking like I(a+b) in a matrix/data.frame/list.
}
\examples{
data <- matrix(c(1:30),ncol=3); colnames(data) <- c("a","b","I(a+b)")
calc.I.cols(data)
calc.I.cols(as.list(as.data.frame(data)))
}
\author{
Daniel Hoop
}
\keyword{internal}
