% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{categ.to.bin}
\alias{categ.to.bin}
\title{Convert categorial data into binomial data.}
\usage{
categ.to.bin(x, varname = "var", sep.sign = "_", allnames = NULL)
}
\description{
Convert categorial data into binomial data.
}
\examples{
x <- sample(1:5, 10, replace=TRUE)
categ.to.bin(x, varname = "Typ", sep.sign = "")
categ.to.bin(x, allnames = paste0("Value_", LETTERS[1:5]))
}
\seealso{
\code{\link{categ.to.ordinal}}
}
\author{
Daniel Hoop
}
