% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{categ.to.ordinal}
\alias{categ.to.ordinal}
\title{Convert categorial data into ordinal data.}
\usage{
categ.to.ordinal(x, varname = "var", sep.sign = "_", allnames = NULL)
}
\description{
Convert categorial data into ordinal data.
}
\examples{
x <- sample(1:5, 10, replace=TRUE)
categ.to.ordinal(x, varname = "Region", sep.sign = "")
}
\seealso{
\code{\link{categ.to.bin}}
}
\author{
Daniel Hoop
}
