% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{cbind.1b1}
\alias{cbind.1b1}
\title{Binds all given matrices in ... column by column.}
\usage{
\method{cbind}{`1b1`}(...)
}
\description{
Binds all given matrices in ... column by column.
}
\seealso{
\code{\link{cbind.1b1}}
}
\author{
Daniel Hoop
}
