% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{ch1903.to.wgs84}
\alias{ch1903.to.wgs84}
\title{Convert from Swiss geographic system (CH1903; y, x) to internation geographic system (WGS84; N, E)}
\usage{
ch1903.to.wgs84(x, y, output = c("decimal", "minsec"))
}
\arguments{
\item{x}{The x coordinates}

\item{y}{The y coordinates}

\item{output}{A character value describing the output format. Eiter \code{"decimal"} (decimal value of length 1) or \code{"minsec"} (vector of length 3).}
}
\value{
The coordinates (N, E) in the internation geographic system (WGS84).
}
\description{
Convert from Swiss geographic system (CH1903; y, x) to internation geographic system (WGS84; N, E)
}
\details{
Note that the coordinates METER_X and METER_Y in the AGIS data are probably interchanged.
}
\examples{
ch1903.to.wgs84(x = 100000, y = 700000, output = "minsec")
# N = 46° 02' 38.86"
# E =  8° 43' 49.80"
}
