% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{changeFileEncoding}
\alias{changeFileEncoding}
\title{Change encoding of a file}
\usage{
changeFileEncoding(
  fileIn,
  fileOut,
  encodingIn = "CP1252",
  encodingOut = "UTF-8",
  verbose = FALSE
)
}
\description{
Change encoding of a file
}
\details{
This function works only on UNIX systems and is based on the 'iconvs' command.\cr
Available encodings are, amongst others,\cr
\itemize{
\item WINDOWS-1252 (or CP1252)
\item ISO-8559-1
\item UTF-8
}
For more encodings, consult: \code{man 7 charsets}
}
\examples{
con <- file("enctest.txt", encoding = "utf-8")
writeLines("a ä o ö u ü", con)
close(con)

changeFileEncoding(fileIn = "enctest.txt", fileOut = "enctest_CP1252.txt", encodingIn = "UTF-8", encodingOut = "WINDOWS-1252")
changeFileEncoding(fileIn = "enctest_CP1252.txt", fileOut = "enctest_UTF-8.txt", encodingIn = "WINDOWS-1252", encodingOut = "UTF-8")

print(readLines("enctest.txt")) # OK.
print(readLines("enctest_CP1252.txt")) # Not looking good...
print(readLines("enctest_UTF-8.txt")) # OK.

file.remove(c("enctest.txt", "enctest_CP1252.txt", "enctest_UTF-8.txt"))
}
\author{
Daniel Hoop
}
