% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{cor.table}
\alias{cor.table}
\title{Calculate correlation talbe and show signification signs.}
\usage{
cor.table(
  data,
  sig.level = c(0.1, 0.05, 0.01, 0.001)[2],
  method = c("pearson", "kendall", "spearman"),
  conf.level = NULL,
  conf.middle = TRUE,
  conf.middle.sign = TRUE,
  digits = 2,
  triangle = FALSE,
  del.diag = TRUE,
  ignore.nonsig = FALSE,
  suppressWarnings = FALSE,
  count = FALSE,
  special.sign.before.nonsig.for.excel = FALSE,
  ...
)
}
\description{
Calculate correlation talbe and show signification signs.
}
\author{
Daniel Hoop
}
