% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{create.cols}
\alias{create.cols}
\title{Create columns in \code{data} that are not already existent.}
\usage{
create.cols(data, cols, non.I.value = NA_integer_)
}
\arguments{
\item{data}{The \code{data.frame} to which columns should be added.}

\item{cols}{Character vector of columns that should be created if the don't exist in \code{data}.}

\item{non.I.value}{The value that is filled into columns that don't have colnames I().}
}
\value{
The same \code{data.frame} as given in argument \code{data} but with the additional columns that were specified in \code{cols}.
}
\description{
Create columns in \code{data} that are not already existent.
}
\details{
If cols like \code{"I(a/b)"} are given, then the function will look for a column called \code{"a"} and a column calles \code{"b"} in \code{data}.
The calculation will then be done as specified, and be written into the according column.
}
\examples{
data <- data.frame(a = 1:3, b = 1:3)
create.cols(data, c("I(a/b)", "I(a+b)", "c"))
}
\author{
Daniel Hoop
}
\keyword{internal}
