% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{create.spb.340.size}
\alias{create.spb.340.size}
\title{Creates pseudo size cols in P340 as GVE taken from P330.}
\usage{
create.spb.340.size(data, sizeCol = "size")
}
\arguments{
\item{data}{The data.frame to perform manipulations in.}

\item{sizeCol}{The column for the size in P340. If \code{sizeCol = "size"}, then, e.g. the following Merkmal would be build for the dairy enterprise: P340_size_11000}
}
\value{
The manipulated data.frame.
}
\description{
Creates pseudo size cols in P340 as GVE taken from P330.
}
\examples{
spb <- load.spb()
spb <- create.spb.340.size(spb)
tail(colnames(spb))
}
\author{
Daniel Hoop
}
