% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{csv.to.rdata}
\alias{csv.to.rdata}
\title{Convert a CSV file to an RData file that contains the same data.}
\usage{
csv.to.rdata(
  path,
  dat = NULL,
  name = "dat",
  clean.numb.format = FALSE,
  clean.dat.columns = FALSE,
  update.csv = FALSE,
  ...
)
}
\description{
Convert a CSV file to an RData file that contains the same data.
}
\author{
Daniel Hoop
}
