% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{dayOfMonth}
\alias{dayOfMonth}
\title{Extract the day of month from a POSIX time object.}
\usage{
dayOfMonth(time)
}
\arguments{
\item{time}{A POSIX time object.}
}
\value{
The dayOfMonth as numeric.
}
\description{
Extract the day of month from a POSIX time object.
}
\examples{
dayOfMonth(Sys.time())
}
\author{
Daniel Hoop
}
