% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{dayOfWeek}
\alias{dayOfWeek}
\title{Extract the day of week from a POSIX time object.}
\usage{
dayOfWeek(time, numeric = TRUE)
}
\arguments{
\item{time}{A POSIX time object.}

\item{numeric}{Logical value indicating if the return value should be numeric (instead of character). Character output will depend on the locale.}
}
\value{
The dayOfWeek as numeric starting with 1 = Monday, ending with 7 = Sunday. Otherwise a character string depending on the locale.
}
\description{
Extract the day of week from a POSIX time object.
}
\examples{
dayOfWeek(Sys.time(), numeric = TRUE)
}
\author{
Daniel Hoop
}
