% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{diffInDataset}
\alias{diffInDataset}
\title{Helper function for \code{\link{assertDatasetEquality}}.}
\usage{
diffInDataset(dat1, dat2, colsToCheck = NULL, tolerance = 1e-09)
}
\description{
Helper function for \code{\link{assertDatasetEquality}}.
}
\author{
Daniel Hoop
}
\keyword{internal}
