% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{dim3.to.mat}
\alias{dim3.to.mat}
\title{Converts an array with 3 dimension to a matrix (with 2 dimensions).}
\usage{
dim3.to.mat(
  x,
  sep.line = TRUE,
  sep.sign = NA,
  keep.colnames = TRUE,
  keep.dim3names = TRUE
)
}
\arguments{
\item{x}{array to be converted. (array with 3 dimensions)}

\item{sep.line}{Should the sub-matrices be separated by the sep.sign? (logical)}

\item{sep.sign}{The sign to separate the sub-matrices (character)}

\item{keep.colnames}{Should the colnames be written into the result matrix as pseudo colnames above each sub-matrix? (logical)}

\item{keep.dim3names}{Should the dimnames of the 3rd dimension be written into the rownames of the result matrix? (logical)}
}
\description{
This is especially useful if you want to export a 3 dimensional array into a csv file.
The matrix shows the third dimension of the original array row by row as 'sub-matrices'.
}
\examples{
x <- array(0, dim=c(5,5,2), dimnames=list(c("asdf1","asdf2","asdf3","asdf4","asdf5"),c("asdf1","asdf2","asdf3","asdf4","asdf5"),c("dim3.1", "dim3.2")))
dim3.to.mat(x, sep.line=TRUE, sep.sign=NA, keep.colnames=TRUE, keep.dim3names=TRUE)
dim3.to.mat(list(a=matrix(1:10, ncol=2), b=matrix(11:20, ncol=2), NULL, c=matrix(21:30, ncol=2), NULL, a=matrix(31:40, ncol=2)))
}
\author{
Daniel Hoop
}
