% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{dimnames.to.mat}
\alias{dimnames.to.mat}
\title{Integrate dimnames to a matrix.}
\usage{
\method{dimnames}{to.mat}(x, dims = c("rowcol", "row", "col"))
}
\arguments{
\item{x}{The matrix into which the dimnames should be integrated.}

\item{dimns}{Character specifying which dimnaes should be integrated. Either \code{"rowcol"}, meaning both, \code{"row"}, meaning rownames, or \code{"col"}, meaning colnames only.}
}
\description{
Integrate dimnames to a matrix.
}
\details{
This function is needed for the function \code{\link{c.matrices}}
}
\examples{
x <- data.frame(testCol=1:10)
y <- data.frame(testCol2=1:5)
dimnames.to.mat(x, dims="cpl")
c.matrices(x, y, integrate.dimnames="col")
}
\author{
Daniel Hoop
}
