% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{.addColsBeforeAfter}
\alias{.addColsBeforeAfter}
\title{Internal function called by \code{\link{addColsAfter}} or \code{\link{addColsBefore}}}
\usage{
.addColsBeforeAfter(
  referenceName,
  addNames = NULL,
  values = NULL,
  data,
  before = TRUE
)
}
\description{
Internal function called by \code{\link{addColsAfter}} or \code{\link{addColsBefore}}
}
\keyword{internal}
