% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{.agsMachineType}
\alias{.agsMachineType}
\title{Show on which Agroscope machine type the R session operates.}
\usage{
.agsMachineType(
  fola = FALSE,
  rstudioserver = FALSE,
  laborpc = FALSE,
  bitpc = FALSE,
  other = FALSE
)
}
\value{
A logical vector with names \code{'rstudioserver'}, \code{'laborpc'}, \code{'fola'}, \code{'bitpc'} and \code{'other'}.
\code{'fola'} is TRUE, when the username starts with an 'f'. A 'laborpc' can be in FOLA or outside FOLA. 'rstudioserver' is always in FOLA.
}
\description{
Show on which Agroscope machine type the R session operates.
}
\details{
All parameters are optional. If one is set to TRUE, then only one logical value is returned. Else a logical vector (see Value).
}
\examples{
if (.agsMachineType()["bitpc"]) print("on BIT client") else print ("on other machine.")
if (.agsMachineType(fola=1)) print("Inside FOLA.")
}
\author{
Daniel Hoop
}
\keyword{internal}
