% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{.checkMissingICols}
\alias{.checkMissingICols}
\title{This function is internally used in \code{\link{mean.weight}}.}
\usage{
.checkMissingICols(colsToCheck, colsAvailable)
}
\description{
This function is internally used in \code{\link{mean.weight}}.
}
\author{
Daniel Hoop
}
\keyword{internal}
