% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{.install.package}
\alias{.install.package}
\title{Install a package if not yet installed.}
\usage{
.install.package(x, repos = options()$repos)
}
\description{
Install a package if not yet installed.
}
\keyword{internal}
