% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{.isNecessaryToUseTmpDir}
\alias{.isNecessaryToUseTmpDir}
\title{Gives the answer to the question if it is necessary to write x to a temporary dir and move it afterwards.}
\usage{
.isNecessaryToUseTmpDir(x, file, sizeThreshold = 1400000)
}
\arguments{
\item{x}{The object to be written to disk. E.g. a matrix or data.frame.}

\item{file}{The file to which the object should be written.}

\item{sizeThreshold}{The threshold of the object.size(x) in bits. If the object is greater than that amount & file is on network drive, temp dir will be used first.}
}
\description{
Function is e.g. necessary for write.table.zipped & write.table (fast version)
}
\author{
Daniel Hoop
}
\keyword{internal}
