% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{.MKtabColRow}
\alias{.MKtabColRow}
\title{Extrahiert aus \code{character} vector oder aus colnames eines Objekts einen beliebigen Abschnitt wie Tabellen-/Spalten-/Zeilenbezeichnung. Gibt \code{character} vector zurueck.}
\usage{
.MKtabColRow(x, which)
}
\arguments{
\item{x}{The character vector or matrix/data.frame}

\item{which}{1 = table, 2 = column, 3 = row}
}
\description{
Extrahiert aus \code{character} vector oder aus colnames eines Objekts einen beliebigen Abschnitt wie Tabellen-/Spalten-/Zeilenbezeichnung. Gibt \code{character} vector zurueck.
}
\author{
Daniel Hoop
}
\keyword{internal}
