% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{.paste.elements}
\alias{.paste.elements}
\title{Helper function for \link{balanced.panel}}
\usage{
.paste.elements(
  l,
  sep = "_",
  errorMsg = "All list places must have same length!"
)
}
\description{
Helper function for \link{balanced.panel}
}
\author{
Daniel Hoop
}
\keyword{internal}
