% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{.prepare.fixed.index.result}
\alias{.prepare.fixed.index.result}
\title{This function is internally used in \code{\link{mean.weight}}.}
\usage{
.prepare.fixed.index.result(
  data,
  index,
  names.result,
  index.sep = "_",
  edit.I.colnames = FALSE
)
}
\description{
This function is internally used in \code{\link{mean.weight}}.
}
\author{
Daniel Hoop
}
\keyword{internal}
