% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{.removeLocalZaData}
\alias{.removeLocalZaData}
\title{Remove the locally stored data.}
\usage{
.removeLocalZaData()
}
\value{
A logical vector incating which files were deleted.
}
\description{
Remove the locally stored data.
}
\author{
Daniel Hoop
}
