% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{.rm.I.from.names}
\alias{.rm.I.from.names}
\title{This function is internally used in \code{\link{mean.weight}}.}
\usage{
.rm.I.from.names(x)
}
\description{
This function is internally used in \code{\link{mean.weight}}.
}
\author{
Daniel Hoop
}
\keyword{internal}
