% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{equal.length}
\alias{equal.length}
\title{Make all entires in vector the same length.}
\usage{
equal.length(
  x,
  add = 0,
  where = c("beginning", "end"),
  minlength = 0,
  margin = 2
)
}
\arguments{
\item{add}{The sign that will be added.}

\item{where}{Character indigating if `add` should be added at the beginning or at the end of the string.}
}
\description{
Make all entires in vector the same length.
}
\examples{
x <- data.frame(a=seq(0.1, 1, 0.1) , b=c(1:10))
equal.length(equal.n.decimals(x))
}
\author{
Daniel Hoop
}
\keyword{internal}
