% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{escapeStr}
\alias{escapeStr}
\title{Escapes characters that have special meaning in regular expressions with backslashes.}
\usage{
escapeStr(x)
}
\arguments{
\item{x}{Character vector containing the strings that should be escaped.}
}
\description{
Escapes characters that have special meaning in regular expressions with backslashes.
}
\examples{
escapeStr(c("asdf.asdf", ".asdf", "\\\\.asdf", "asdf\\\\.asdf") )
}
\author{
Daniel Hoop
}
\keyword{internal}
