% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{extract.I.vars}
\alias{extract.I.vars}
\title{Extracts all variables in a \code{I(a+b*c)} formula seperately. This is useful in combination with the function \code{\link{mean.weight}}.}
\usage{
extract.I.vars(
  vars,
  keep.original = FALSE,
  keep.only.necessary = TRUE,
  original.single.vars.only = FALSE
)
}
\description{
Extracts all variables in a \code{I(a+b*c)} formula seperately. This is useful in combination with the function \code{\link{mean.weight}}.
}
\author{
Daniel Hoop
}
\keyword{internal}
