% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{file.copy}
\alias{file.copy}
\title{Fast version of file.copy implemented by \code{\link[fs:file_copy]{fs::file_copy}}}
\usage{
file.copy(
  from,
  to,
  overwrite = recursive,
  recursive = FALSE,
  copy.mode = TRUE,
  copy.date = TRUE
)
}
\description{
However, it behaves the same as does \code{\link[base:file.copy]{base::file.copy}}. The default arguments are set to match \code{\link[fs:file_copy]{fs::file_copy}} behaviour.
If any of \code{recursive}, \code{copy.mode} or '\code{copy.date} are not set to default, then \code{\link[utils::file.copy]{utils::file.copy}} will be called (not \code{\link[fs:file_copy]{fs::file_copy}}).
}
\seealso{
\code{\link[fs:file_copy]{fs::file_copy}}, \code{\link[base:file.copy]{base::file.copy}}.
}
\author{
Daniel Hoop
}
