% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{file.copy.safely}
\alias{file.copy.safely}
\title{Copy large files safely}
\usage{
file.copy.safely(
  from,
  to,
  overwrite = FALSE,
  copy.mode = TRUE,
  copy.date = TRUE
)
}
\description{
First creates a temporary filename. Upon successfull copy, renames the file
}
\author{
Daniel Hoop
}
