% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{file.move.safely}
\alias{file.move.safely}
\title{Moves large files safely}
\usage{
file.move.safely(from, to, overwrite = FALSE)
}
\description{
First creates a temporary filename. Upon successfull move, renames the file.
}
\author{
Daniel Hoop
}
