% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{find.col}
\alias{find.col}
\title{Find a column in a data.frame using a pattern.}
\usage{
find.col(pattern, dat, ignore.case = TRUE, ...)
}
\description{
Find a column in a data.frame using a pattern.
}
\author{
Daniel Hoop
}
