% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{find.collinear.variables.det.cov}
\alias{find.collinear.variables.det.cov}
\title{Find collinear with determinant of the covariance matrix.}
\usage{
find.collinear.variables.det.cov(
  formula = NULL,
  data,
  xVars = NULL,
  giveThresholdInfo = TRUE
)
}
\arguments{
\item{formula}{A formula object. Must be given, if \code{xVars} is not specified.}

\item{data}{data.frame that contains the variables (i.e. environment in which the created formula should be evaluated).}

\item{xVars}{Character vector containing the names of independent variables to be checked. Must be given, if \code{formula} is not specified.}

\item{giveThresholdInfo}{Logical value indicating if information on the threshold should be given as a message in the console.}
}
\description{
The determinant of the covariance matrix ranges from \strong{0 (Perfect Collinearity) to 1 (No Collinearity)}.
}
\references{
\href{http://stackoverflow.com/questions/3042117/screening-multicollinearity-in-a-regression-model}{http://stackoverflow.com/questions/3042117/screening-multicollinearity-in-a-regression-model}
}
\author{
Daniel Hoop
}
