% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{find.collinear.variables.kappa}
\alias{find.collinear.variables.kappa}
\title{Find collinear using "Kappa".}
\usage{
find.collinear.variables.kappa(
  formula = NULL,
  data,
  xVars = NULL,
  giveThresholdInfo = TRUE
)
}
\arguments{
\item{formula}{A formula object. Must be given, if \code{xVars} is not specified.}

\item{data}{data.frame that contains the variables (i.e. environment in which the created formula should be evaluated).}

\item{xVars}{Character vector containing the names of independent variables to be checked. Must be given, if \code{formula} is not specified.}

\item{giveThresholdInfo}{Logical value indicating if information on the threshold should be given as a message in the console.}
}
\description{
\strong{A Condition Number (CN) > 30 indicates collinearity.}.
}
\references{
\href{http://stackoverflow.com/questions/3042117/screening-multicollinearity-in-a-regression-model}{http://stackoverflow.com/questions/3042117/screening-multicollinearity-in-a-regression-model}, \code{\link[base:kappa]{base::kappa}}
}
\author{
Daniel Hoop
}
