% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{find.collinear.variables.rsq}
\alias{find.collinear.variables.rsq}
\title{Find collinear using R squared.}
\usage{
find.collinear.variables.rsq(
  formula = NULL,
  data,
  xVars = NULL,
  giveThresholdInfo = TRUE
)
}
\arguments{
\item{formula}{A formula object. Must be given, if \code{xVars} is not specified.}

\item{data}{data.frame that contains the variables (i.e. environment in which the created formula should be evaluated).}

\item{xVars}{Character vector containing the names of independent variables to be checked. Must be given, if \code{formula} is not specified.}

\item{giveThresholdInfo}{Logical value indicating if information on the threshold should be given as a message in the console.}
}
\description{
A regression of x_i on all other predictors gives R^2_i. Repeat for all predictors. \strong{R^2_i close to one indicates a problem - the offending linear combination may be found}.
}
\references{
\href{http://stackoverflow.com/questions/3042117/screening-multicollinearity-in-a-regression-model}{http://stackoverflow.com/questions/3042117/screening-multicollinearity-in-a-regression-model}
}
\author{
Daniel Hoop
}
