% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{find.collinear.variables.vif}
\alias{find.collinear.variables.vif}
\title{Find collinear using the variance inflation factor (VIF).}
\usage{
find.collinear.variables.vif(
  formula = NULL,
  data = NULL,
  model = NULL,
  giveThresholdInfo = TRUE
)
}
\arguments{
\item{formula}{A formula. Necessary, only if \code{model} was not given.}

\item{data}{The data to fit the linear model specified in \code{formula}. Necessary, only if \code{model} was not given.}

\item{model}{A fitted model. If given, then \code{formula} and \code{data} don't have to be specified.}

\item{giveThresholdInfo}{Logical value indicating if information on the threshold should be given as a message in the console.}
}
\description{
\strong{VIF > 5 is of concern}. In that case, parameter estimates and probability values are questionable. However, read the details section!
}
\details{
\href{https://statisticsbyjim.com/regression/multicollinearity-in-regression-analysis/}{https://statisticsbyjim.com/regression/multicollinearity-in-regression-analysis/}\cr
VIFs start at 1 and have no upper limit. A value of 1 indicates that there is no correlation between this independent variable and any others. VIFs between 1 and 5 suggest that there is a moderate correlation, but it is not severe enough to warrant corrective measures. VIFs greater than 5 represent critical levels of multicollinearity where the coefficients are poorly estimated, and the p-values are questionable.\cr
\href{https://statisticalhorizons.com/multicollinearity}{https://statisticalhorizons.com/multicollinearity}\cr
Personally, I tend to get concerned when a VIF is greater than 2.50, which corresponds to an R2 of .60 with the other variables.\cr
Regardless of your criterion for what constitutes a high VIF, there are at least three situations in which a high VIF is not a problem and can be safely ignored:\cr
1. The variables with high VIFs are control variables, and the variables of interest do not have high VIFs.\cr
2. The high VIFs are caused by the inclusion of powers or products of other variables.\cr
3. The variables with high VIFs are indicator (dummy) variables that represent a categorical variable with three or more categories.
Futher notes...
VIFs exceeding 4 warrant further investigation, while VIFs exceeding 10 are signs of serious multicollinearity requiring correction.
Tolerance is the reciprocal of VIF.
}
\author{
Daniel Hoop
}
