% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{getXmlNodeIndexByNameAndAttr}
\alias{getXmlNodeIndexByNameAndAttr}
\title{Get the index of a xml node by providing the name and attribute of that node.}
\usage{
getXmlNodeIndexByNameAndAttr(xml, name, attr)
}
\arguments{
\item{xml}{The xml root node (derived by XML::xmlRoot(XML::xmlParse(filename)))}

\item{name}{The name of the node that shall be searched. length(name) must be 1.}

\item{attr}{The attributes of the node that shall be searched. length(attr) can be greater than 1.}
}
\value{
A numeric vector with the indexes that match the given search criteria.
}
\description{
Get the index of a xml node by providing the name and attribute of that node.
}
