% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{group.by.fix.scale}
\alias{group.by.fix.scale}
\title{Group elements of a vector according to given absolute scale.}
\usage{
group.by.fix.scale(
  x,
  selection.levels,
  method = c("< x <=", "<= x <"),
  include.min.max = FALSE,
  give.names = FALSE,
  names.digits = 2,
  names.sep = "-",
  equal.length.names = FALSE
)
}
\arguments{
\item{x}{The vector containing the values relevant for the grouping.}

\item{selection.levels}{The levels to seggregate the data in \code{x}.}

\item{method}{\code{"< x <="} means that the the smallest value of the group is stricly above the lower group boundary.\cr
\code{"<= x <"} means that the smallest value of a group can be the same as the lower group boundary.}

\item{include.min.max}{Logical value indicating if the the min and max should be included in the groups, even if they may be exactly the same as the group boundary level.}

\item{give.names}{Logical value indicating if the grouping vector should have names. The names will show the group boundaries.}

\item{names.digits}{The digits to ground the grouping names if \code{give.names = TRUE}.}

\item{names.sep}{The group boundary separation sign for grouping names if \code{give.names = TRUE}.}

\item{equal.length.names}{Logical value indicating if the names should be striclty of equal length, e.g. \code{"1-10"} would be changed to \code{" 1-10"}.}
}
\description{
Group elements of a vector according to given absolute scale.
}
\details{
If \code{include.min.max=TRUE}, then the outermost observations are also included.
Otherwise, if you give exactely the min(x) and max(x), one of both would be excluded.
If `selection.levels` is only one value, then all values below will be numbered 1 and all values above will be numbered 2.
}
\examples{
group.by.fix.scale(c(0.2,1,10,1000), c(0.1,9.3,20,2000), give.names=TRUE, equal.length.names=TRUE)
}
\seealso{
\code{\link[base:cut]{base::cut}}
}
\author{
Daniel Hoop
}
