% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{group.by.quantiles}
\alias{group.by.quantiles}
\title{Group elements of a vector by quantiles.}
\usage{
group.by.quantiles(
  x,
  probs = seq(0, 1, 0.1),
  method = c("< x <=", "<= x <"),
  index = NULL,
  include.min.max = TRUE,
  give.names = FALSE,
  names.digits = 2,
  names.sep = "-",
  equal.length.names = FALSE,
  weights = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{x}{The vector containing the values relevant for the grouping.}

\item{probs}{The quantile probabilities to segregate the vector given in \code{x}.}

\item{method}{\code{"< x <="} means that the the smallest value of the group is stricly above the lower group boundary.\cr
\code{"<= x <"} means that the smallest value of a group can be the same as the lower group boundary.}

\item{include.min.max}{Logical value indicating if the the min and max should be included in the groups, even if they may be exactly the same as the group boundary level.}

\item{give.names}{Logical value indicating if the grouping vector should have names. The names will show the group boundaries.}

\item{names.digits}{The digits to ground the grouping names if \code{give.names = TRUE}.}

\item{names.sep}{The group boundary separation sign for grouping names if \code{give.names = TRUE}.}

\item{equal.length.names}{Logical value indicating if the names should be striclty of equal length, e.g. \code{"1-10"} would be changed to \code{" 1-10"}.}
}
\description{
This is a wrapper for \code{\link{group.by.fix.scale}} with a slightly altered interface.
}
\examples{
data <- data.frame(x = rnorm(100), weights = rnorm(100, 10))
data[,"gr"] <- group.by.quantiles(x = data[,"x"], weights = data[,"weights"])
head(data)
}
\seealso{
\code{\link{group.by.fix.scale}}
}
\author{
Daniel Hoop
}
