% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{group.by.quartiles}
\alias{group.by.quartiles}
\title{Group data by quantiles.}
\usage{
group.by.quartiles(...)
}
\description{
Like \code{\link{group.by.quantiles}}, but with other default argument `probs`.
}
\examples{
data <- data.frame(x = rnorm(100), weights = rnorm(100, 10))
data[,"gr"] <- group.by.quantiles(x = data[,"x"], weights = data[,"weights"])
head(data)
}
\seealso{
\code{\link{group.by.quantiles}}
}
\author{
Daniel Hoop
}
