% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{group.by.wtd.quantiles}
\alias{group.by.wtd.quantiles}
\title{Group elements of a vector by weighted quantiles.}
\usage{
group.by.wtd.quantiles(
  x,
  weights = NULL,
  index = NULL,
  probs = c(0, 0.25, 0.5, 0.75, 1),
  method = c("< x <=", "<= x <"),
  na.rm = TRUE,
  include.zero.weight.obs = FALSE
)
}
\arguments{
\item{x}{The vector containing the values relevant for the grouping.}

\item{weights}{Either NULL or a numeric vector giving the weights.}

\item{probs}{The quantile probabilities to segregate the vector given in \code{x}.}

\item{method}{\code{"< x <="} means that the the smallest value of the group is stricly above the lower group boundary.\cr
\code{"<= x <"} means that the smallest value of a group can be the same as the lower group boundary.}

\item{include.zero.weight.obs}{Logical value indicating if observations with weight 0 should be included into the outer groups in case they had not beed allocated to them.}
}
\description{
Like \code{\link{group.by.quantiles}}, but weighted.
}
\examples{
data <- data.frame(x = rnorm(100), weights = rnorm(100, 10))
data[,"gr"] <- group.by.wtd.quantiles(x = data[,"x"], weights = data[,"weights"])
head(data)
}
\seealso{
\code{\link{group.by.quartiles}}
}
\author{
Daniel Hoop
}
