% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{hour}
\alias{hour}
\title{Extract the hour from a POSIX time object.}
\usage{
hour(time)
}
\arguments{
\item{time}{A POSIX time object.}
}
\value{
The hour as numeric.
}
\description{
Extract the hour from a POSIX time object.
}
\examples{
hour(Sys.time())
}
\author{
Daniel Hoop
}
