% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{id.entschluesseln}
\alias{id.entschluesseln}
\title{Decode encrypted IDs of Referenzbetriebe.}
\usage{
id.entschluesseln(...)
}
\description{
Decode encrypted IDs of Referenzbetriebe.
}
\author{
Daniel Hoop
}
\keyword{internal}
