% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{installFromCRAN}
\alias{installFromCRAN}
\title{Install a package from CRAN, but only if it is not yet installed}
\usage{
installFromCRAN(pkgs, lib = NULL, repos = options()$repos, ...)
}
\arguments{
\item{lib}{character vector giving the library directories where to install the packages. Recycled as needed. If missing, defaults to the first element of \code{\link[base:.libPaths]{base::.libPaths}()}.}

\item{...}{Further arguments passed into \code{\link[utils:install.packages]{utils::install.packages}}.}

\item{x}{The name(s) of the package(s).}
}
\value{
Logical vector (always FALSE) with length(pkgs), i.e. `return(logical(length(pkgs)))`. For compability with other installation methods of the package.
}
\description{
Install a package from CRAN, but only if it is not yet installed
}
\author{
Daniel Hoop
}
\keyword{internal}
