% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{isMM}
\alias{isMM}
\title{Determine if something looks like a Merkmal (e.g. "P300_2300_30000")}
\usage{
isMM(x, strict = FALSE)
}
\arguments{
\item{x}{Either a data.frame/matrix or a character vector. If a data.frame/matrix is given, then the information is given with respect to the colnames.}

\item{strict}{Logical value indicating if all identifiers have to be numeric. If \code{strict = TRUE}, then \code{'P340_2399_3000'} would be valid, but \code{'P320_abcd_40000'} would be not. If \code{strict = FALSE}, then also \code{'Pxxx_woidjfoiwe_dcnv823'} is a valid Merkmal name.}
}
\value{
A logical vector indicating for each element given in \code{x} if it is a Merkmal or not.
}
\description{
Determine if something looks like a Merkmal (e.g. "P300_2300_30000")
}
\details{
Names starting with \code{"M"} or with \code{"P"} are allowed. As delimiters, \code{"."} and \code{"_"} are possible. E.g. \code{"P300_2000_30000"}, or \code{"M300.200.30000"} are both valid names.
}
\examples{
isMM(c("M230_3000_200aa", "X230_3900_20000"))
# [1] TRUE FALSE

isMM(data.frame(M230_3000_200aa = 1, P39.2300.200 = 1), strict = TRUE)
# [1] FALSE TRUE
}
\author{
Daniel Hoop
}
