% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{kruskal.multiple}
\alias{kruskal.multiple}
\title{Conducts zaUtils::kruskal.groups agricolae::kruskal for multiple columns in a data.frame.}
\usage{
kruskal.multiple(
  data,
  grouping,
  sig.level = 0.05,
  p.adj = "holm",
  group = TRUE,
  ...
)
}
\arguments{
\item{data}{The data.frame containing the data to test.}

\item{grouping}{A grouping vector.}

\item{sig.level}{The significance level which is relevant for the grouping with letters.}

\item{p.adj}{The probability adjustment method.}

\item{group}{Logical value indicating if a grouping with letters should be conducted (`zaUtils::kruskal.groups`) or not (`agricolae::kruskal`).}
}
\description{
Conducts zaUtils::kruskal.groups agricolae::kruskal for multiple columns in a data.frame.
}
\author{
Daniel Hoop
}
