% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{kruskal.test.multiple}
\alias{kruskal.test.multiple}
\title{Conducts stats::kruskal.test for multiple columns in a data.frame.}
\usage{
\method{kruskal.test}{multiple}(data, grouping, extract.p = TRUE, digits = 3)
}
\arguments{
\item{data}{The data.frame containing the data to test.}

\item{grouping}{A grouping vector.}

\item{extract.p}{Logical value if only the p value should be extracted from the result of `stats::kruskal.test`.}

\item{digits}{The number of digits to round the p values.}
}
\description{
Conducts stats::kruskal.test for multiple columns in a data.frame.
}
\examples{
kruskal.test.multiple(
  data.frame(a = 1:10, b = runif(10)),
  c(rep(1, 5), rep(2, 5)))
}
\author{
Daniel Hoop
}
