% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{kruskalmc.multiple}
\alias{kruskalmc.multiple}
\title{Conducts pgirmess::kruskalmc for multiple columns in a data.frame.}
\usage{
kruskalmc.multiple(data, grouping, sig.level = 0.05)
}
\arguments{
\item{data}{The data.frame containing the data to test.}

\item{grouping}{A grouping vector.
kruskalmc.multiple(
  data.frame(a = 1:10, b = runif(10)),
  c(rep(1, 5), rep(2, 5)))}
}
\description{
Conducts pgirmess::kruskalmc for multiple columns in a data.frame.
}
\author{
Daniel Hoop
}
