% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{left.to.right}
\alias{left.to.right}
\title{Flips matrix left to right  (same as \code{\link{right.to.left}}).}
\usage{
left.to.right(x)
}
\description{
Flips matrix left to right  (same as \code{\link{right.to.left}}).
}
\author{
Daniel Hoop
}
\keyword{internal}
