% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{load.agis}
\alias{load.agis}
\title{Load AGIS data}
\usage{
load.agis(
  year,
  additional = FALSE,
  status = FALSE,
  keepAllInStatus = FALSE,
  cols = NULL,
  filterFunc = NULL
)
}
\arguments{
\item{year}{A vector containing all years for which AGIS data should be loaded.}

\item{additional}{Logical value indicating if addional data such as "Auswahlgesamtheit" should be loaded as well.}

\item{status}{Logical value indicating if additional data from the recruitment status should be loaded as well.}

\item{keepAllInStatus}{Logical value indicating if all farms in the 'status' dataset should be kept. If this option is set TRUE, then the number of farms in 'status' dataset and in AGIS dataset must be equal.}

\item{cols}{Optional: Character vector listing the columns that should be extracted. The rest of the columns will be discarded.\cr
Special cases: If \code{NULL}, then all columns will be loaded. If \code{"Rueckverfolg"}, then columns that are relevant for the "Rueckverfolgung" are loaded. Else, just provide the columns you want to load.}

\item{filterFunc}{Optional: A function to filter the dataset while reading the data from all year. See the example for more details.}
}
\description{
Load AGIS data
}
\examples{
agis <- load.agis(2017:2018, filterFunc = function(agis) agis[agis[, "ZA_ID"] \%in\% spe[, "ZA_ID"], ])
}
\author{
Daniel Hoop
}
