% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{load.cost}
\alias{load.cost}
\title{Load full cost data}
\usage{
load.cost(
  years = 2014,
  ignore_P_cols = TRUE,
  non_aggr = FALSE,
  filter_expression = NULL,
  filterFunc = NULL,
  parentDir = NULL
)
}
\description{
Load full cost data
}
\author{
Daniel Hoop
}
