% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{load.gb}
\alias{load.gb}
\title{Load Referenzbetriebs data (2003 - 2015) with the same columns that are given in the Grundlagenbericht.}
\usage{
load.gb(inklBG = FALSE)
}
\description{
Load Referenzbetriebs data (2003 - 2015) with the same columns that are given in the Grundlagenbericht.
}
\details{
In the year 2015 weights are 0 for all farms because SpE was the official sample in that year.
}
\author{
Daniel Hoop
}
