% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{load.spb.gb}
\alias{load.spb.gb}
\title{Load SpB data with the same columns that are given in the Grundlagenbericht.}
\usage{
load.spb.gb()
}
\description{
Load SpB data with the same columns that are given in the Grundlagenbericht.
}
\author{
Daniel Hoop
}
