% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{load.spb}
\alias{load.spb}
\title{Load SpB raw data.}
\usage{
load.spb(withGbVars = FALSE)
}
\arguments{
\item{withGbVars}{Logical value indicating if the variables that are defined for the "Grundlagenbericht" should be extracted as well.}
}
\description{
Load SpB raw data.
}
\author{
Daniel Hoop
}
