% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{load.spe.gb}
\alias{load.spe.gb}
\title{Load SpE data with the same columns that are given in the Grundlagenbericht.}
\usage{
load.spe.gb()
}
\description{
Load SpE data with the same columns that are given in the Grundlagenbericht.
}
\author{
Daniel Hoop
}
