% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{load.spe}
\alias{load.spe}
\title{Load SpE raw data.}
\usage{
load.spe(
  withGbVars = FALSE,
  assureSameObsInDbAndStatusList = FALSE,
  latestBhjForTest = NULL
)
}
\arguments{
\item{withGbVars}{Logical value indicating if the variables that are defined for the "Grundlagenbericht" should be extracted as well.}

\item{assureSameObsInDbAndStatusList}{Logical value indicating if it should be assured that the number of observations in the database is in accordance with the number of observartions in the status list.
The function will take into account the fact that some observations could not be linked with AGIS data, and will not report those missing observations as errors.}

\item{latestBhjForTest}{The latest book keeping year for the above mentioned test.}
}
\description{
Load SpE raw data.
}
\author{
Daniel Hoop
}
