% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{load.spe.status}
\alias{load.spe.status}
\title{Load the latest status file for SpE}
\usage{
load.spe.status(BHJ = NULL, termin = NULL, file = NULL)
}
\arguments{
\item{BHJ}{*Optional*: Buchhaltungsjahr}

\item{termin}{*Optional*: If not given, the latest one will be chosen automatically.}

\item{file}{*Optional*: If `BHJ` is not given, then `file` must be specified.}
}
\description{
Load the latest status file for SpE
}
\author{
Daniel Hoop
}
