% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{load.tbn.gb}
\alias{load.tbn.gb}
\title{Load Testbetriebsnetz data (1990 - 2002) with the same columns that are given in the Grundlagenbericht.}
\usage{
load.tbn.gb(inklBG = FALSE)
}
\description{
Load Testbetriebsnetz data (1990 - 2002) with the same columns that are given in the Grundlagenbericht.
}
\details{
Some columns that are available from \code{\link{load.gb}} are not available from this function because the< were not converted from Testbetriebsnetz- to Referenzbetriebe-Merkmalskatalog and would therefore always be 0. To avoid misinterpretation of these variables they were been dropped from the dataset.
}
\author{
Daniel Hoop
}
