% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{load2}
\alias{load2}
\title{Load an object from an RData file and return it.}
\usage{
load2(file, getObjName = FALSE)
}
\arguments{
\item{file}{a (readable binary-mode) \link[base]{connection} or a character string
    giving the name of the file to load (when \link[base]{tilde expansion}
    is done).}

\item{getObjName}{Logical value indicating if the object name should also be returned.}
}
\value{
If \code{getObjName = FALSE}, the loaded object. Else, a list with two entries "object" (the loaded object), and "name" (the name of the object).
}
\description{
Will only work, if the file contains only 1 object.
}
\author{
Daniel Hoop
}
