% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{long.to.wide.df}
\alias{long.to.wide.df}
\title{Erstellt aus einem langen Datensatz (mit vielen Zeilen), wo alle \code{rawIndex} untereinander stehen und indexiert sind, einen breiten \code{data.frame}, wo die indexierten Daten stattdessen in verschiedenen Spalten stehen.}
\usage{
long.to.wide.df(
  dat,
  colnameFixed = c("ID", "Jahr"),
  colnameIndex = "index",
  colnameValue = "Franken"
)
}
\arguments{
\item{dat}{'Langer' Datensatz in indexierter Form.}

\item{colnameFixed}{Spalten welche die Betriebsinformationen enthalten (z.B. `"ID"` und `"Jahr"`). Kann auch `NULL`` sein. Dann werden keine Zusatzinfos angehaengt.}

\item{colnameIndex}{Name der Index-Spalte in `dat`.}

\item{colnameValue}{Name der Spalte in `dat`, welche die Werte enthaelt.}
}
\description{
Erstellt aus einem langen Datensatz (mit vielen Zeilen), wo alle \code{rawIndex} untereinander stehen und indexiert sind, einen breiten \code{data.frame}, wo die indexierten Daten stattdessen in verschiedenen Spalten stehen.
}
\seealso{
\code{\link[stats:reshape]{stats::reshape}}

\code{\link[reshape:cast]{reshape::cast}}, \code{\link[stats:reshape]{stats::reshape}}, \code{\link{wide.to.long.df}}
}
\author{
Daniel Hoop
}
