% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{match.df.by.id}
\alias{match.df.by.id}
\title{Matches two data frames by id.}
\usage{
match.df.by.id(
  df1,
  df2,
  id1,
  id2,
  keep.no.matches = TRUE,
  check.duplicated = TRUE,
  stringsAsFactors = FALSE
)
}
\arguments{
\item{df1}{The first `data.frame`.}

\item{df2}{The second `data.frame`.}

\item{id1}{A vector of ids for the first `data.frame`.}

\item{id2}{A vector of ids for the second `data.frame`.}

\item{keep.no.matches}{Logical determining whether rows that could not. Default `TRUE`.}

\item{check.duplicated}{Logical determining whether the ids should be checked for duplicated entries in each vector before matching. Default `TRUE`.}

\item{stringsAsFactors}{Logical determining wheter character vectors be converted to factors. Default `FALSE`.}
}
\value{
A `data.frame` containing `i1` and `id2` in the first two columns followed by the columns of `df1` and `df2`.
}
\description{
Matches two data frames by id.
}
\seealso{
\code{\link[base:merge]{base::merge}}
}
\author{
Daniel Hoop
}
