% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{match.multiple.id.left}
\alias{match.multiple.id.left}
\title{Match from uniqe ids on the right side, to multiple ids on the left side.}
\usage{
match.multiple.id.left(id_left, id_right)
}
\arguments{
\item{id_left}{The vector containing the ids in the "left" dataset (into which the data should be filled).}

\item{id_right}{The vector containing the ids in the "right" dataset (from which the data will be drawn).}
}
\description{
Match from uniqe ids on the right side, to multiple ids on the left side.
}
\details{
This is useful if you want to "distribute" data from a smaller dataset to a larger dataset in which the same observations occur several times.
}
\author{
Daniel Hoop
}
