% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{max.n.digits}
\alias{max.n.digits}
\title{Finds the maximal relevant digits for a numeric vector 'x'.}
\usage{
\method{max}{n.digits}(x, startDigits = 5, tol = 10^(-startDigits))
}
\description{
It rounds all numbers from 'startDigits' to 0 and checks if the difference between the original values and the rounded values are smaller than 'tol'.
}
\examples{
max.n.digits(c(0.1, 0.111, 10000))
# [1] 3
}
\author{
Daniel Hoop
}
\keyword{internal}
