% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{mean.weight}
\alias{mean.weight}
\title{Calculates the weighted mean of all variables in a possibly indexed data.frame or matrix.}
\usage{
\method{mean}{weight}(
  data,
  weights = NULL,
  index = NULL,
  cols = NULL,
  fixed.index = FALSE,
  index.of.result = NULL,
  index.sep = "_",
  trunc.probs = NULL,
  calc.sum = FALSE,
  digits = NULL,
  na.rm = TRUE,
  ignore.I = FALSE,
  edit.I.colnames = FALSE,
  del.I.help.columns = FALSE,
  I.help.columns = NULL,
  change.output.str = FALSE
)
}
\arguments{
\item{data}{The data of which the weighted means should be calculated. Can be \code{data.frame}, \code{matrix} or \code{vector}.
If any colname of data contains an expression like \code{I(a/b)}, then the the "ratio of means" is calculated.}

\item{weights}{Optionally: A numeric vector containing the weights for each observation given in \code{x}.}

\item{index}{Optionally: A numeric vector or \code{data.frame}/\code{list} containing the indices. See also \code{\link[base:tapply]{base::tapply}}.}

\item{cols}{Optionally: The columns in data for which the means should be caldulated. If \code{data} contains the columns \code{"a"} and \code{"b"}, then \code{cols} can also be \code{"I(a/b)"}
even though the column \code{"I(a/b)"} is not acually contained in \code{data}.}

\item{trunc.probs}{Optionally: The probability of quantiles within which the truncated mean should be calculated.}

\item{calc.sum}{Should \code{sum(data*weights)} be calculated, rather than weighted means? TRUE/FALSE}

\item{digits}{The number of digits according to which the result will be rounded.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{ignore.I}{Logical value indicating if expressions like \code{I(a/b)} should be ignored in colnames. In that case, *not* "ratio of means" would be calculated, but "mean of ratios".}

\item{edit.I.colnames}{Should the colnames containing expressions with \code{I()} be edited, such that \code{I()} won't be there anymore? TRUE/FALSE}

\item{del.I.help.columns}{Deprecated (use \code{cols} instead): Logical value indicating if the columns provided in \code{I.help.columns} should be removed from the results.}

\item{I.help.columns}{Deprecated (use \code{cols} instead): Character vector giving 'help' columns that are only relevant to calculate other Columns names like \code{I(a/b)} and should be discarded in the final result}

\item{change.output.str}{Logical value indicating if the output structure should be changed in case more than 1 index is given in \code{index}.}
}
\description{
Calculates the weighted mean of all variables in a possibly indexed data.frame or matrix.
}
\details{
If the argument \code{cols} is given, then the result will never be dropped to a vector but always be kept as a matrix, even if only 1 columns was chosen from \code{data}.
}
\examples{
# Prepare input data
data <- as.data.frame(matrix(1:15, ncol=3))
colnames(data) <- c("I(a+b)","a","b")
weights <- 1:5
index <- as.data.frame(matrix(c(2014,2014,2014,2015,2016,   1,2,2,1,1,   11,11,12,13,13),ncol=3))
index.of.result=c("2014_2_11","2014_1_11","0000_0_00")
# Apply function
# By default, all columns in data will be computed
mean.weight(data, weights, index)
# You can also specify the argument `cols`. If cols are given, the
mean.weight(data, weights, index, cols = "I(a/b)")
# Apply to a vector
mean.weight(data[,1], weights, index)
# Apply to a vector, but with fixed fixed.index=TRUE, and specifying the index of the result.
mean.weight(data[,1], weights, index, fixed.index=TRUE, index.of.result=index.of.result)
}
\author{
Daniel Hoop
}
