% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{median.weight}
\alias{median.weight}
\title{Calculate weighted medians.}
\usage{
\method{median}{weight}(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=quantile.weight]{quantile.weight}}
  \describe{
    \item{\code{x}}{The \code{vector}, \code{data.frame}, or \code{matrix} that contains the data for which the quantiles should be calculated.}
    \item{\code{weights}}{Optionally: A numeric vector containing the weights for each observation given in \code{x}.}
    \item{\code{index}}{Optionally: A numeric vector or \code{data.frame}/\code{list} containing the indices. See also \code{\link[base:tapply]{base::tapply}}.}
    \item{\code{probs}}{The probabilities of the quantiles that should be calculated.}
    \item{\code{na.rm}}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
  }}
}
\description{
Calculate weighted medians.
}
\seealso{
\code{\link{quantile.weight}}
}
\author{
Daniel Hoop
}
