% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{merge.gb}
\alias{merge.gb}
\title{Merge Tables A, B, C, D, E, F into one large einzelbetriebliches Referenzbetriebs-GB dataset.}
\usage{
\method{merge}{gb}(
  folder,
  filenames = NULL,
  extra_filename = NULL,
  update.files = FALSE,
  save.file = TRUE,
  save.name = "GB",
  filetype = c("csv"),
  NAto0 = TRUE,
  header = TRUE,
  colnamesrow = 1,
  skiprows = colnamesrow + 1,
  idCol = "ID",
  yearCol = "Jahr",
  ...
)
}
\description{
Merge Tables A, B, C, D, E, F into one large einzelbetriebliches Referenzbetriebs-GB dataset.
}
\author{
Daniel Hoop
}
\keyword{internal}
