% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{merge.matrices}
\alias{merge.matrices}
\title{Deprecated, use \code{\link{c.matrices}} instead}
\usage{
\method{merge}{matrices}(...)
}
\description{
Deprecated, use \code{\link{c.matrices}} instead
}
\author{
Daniel Hoop
}
