% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{minute}
\alias{minute}
\title{Extract the minute from a POSIX time object.}
\usage{
minute(time)
}
\arguments{
\item{time}{A POSIX time object.}
}
\value{
The minute as numeric.
}
\description{
Extract the minute from a POSIX time object.
}
\examples{
minute(Sys.time())
}
\author{
Daniel Hoop
}
