% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{MKcol<-}
\alias{MKcol<-}
\title{Setzt in \code{character} vector, oder in colnames eines Objekts die Nummer der ZA-BH-Spalte ein.}
\usage{
MKcol(x) <- value
}
\description{
Setzt in \code{character} vector, oder in colnames eines Objekts die Nummer der ZA-BH-Spalte ein.
}
\author{
Daniel Hoop
}
