% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{MKcol}
\alias{MKcol}
\title{Extrahiert aus \code{character} vector oder aus colnames eines Objekts die Nummer der ZA-BH-Spalte. Gibt \code{character} vector zurueck.}
\usage{
MKcol(x)
}
\description{
Extrahiert aus \code{character} vector oder aus colnames eines Objekts die Nummer der ZA-BH-Spalte. Gibt \code{character} vector zurueck.
}
\author{
Daniel Hoop
}
