% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{MKrow<-}
\alias{MKrow<-}
\title{Setzt in \code{character} vector, oder in colnames eines Objekts die Nummer der ZA-BH-Zeile ein.}
\usage{
MKrow(x) <- value
}
\description{
Setzt in \code{character} vector, oder in colnames eines Objekts die Nummer der ZA-BH-Zeile ein.
}
\author{
Daniel Hoop
}
