% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{MKrow}
\alias{MKrow}
\title{Extrahiert aus \code{character} vector oder aus colnames eines Objekts die Nummer der ZA-BH-Zeile. Gibt \code{character} vector zurueck.}
\usage{
MKrow(x)
}
\description{
Extrahiert aus \code{character} vector oder aus colnames eines Objekts die Nummer der ZA-BH-Zeile. Gibt \code{character} vector zurueck.
}
\author{
Daniel Hoop
}
