% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{MKsort.colnames}
\alias{MKsort.colnames}
\title{Same as \code{\link{MKsort.colnames}}, but will return the colnames of \code{data} instead of \code{data} itself.}
\usage{
MKsort.colnames(data, order = c("zeile", "spalte"))
}
\description{
Same as \code{\link{MKsort.colnames}}, but will return the colnames of \code{data} instead of \code{data} itself.
}
\author{
Daniel Hoop
}
\keyword{internal}
