% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{MKsort}
\alias{MKsort}
\title{Sort a dataset according to the colnames, where Merkmale like 'P430_0100_94000' are treated specially.}
\usage{
MKsort(data, order = c("row", "col"))
}
\description{
Sort a dataset according to the colnames, where Merkmale like 'P430_0100_94000' are treated specially.
}
\author{
Daniel Hoop
}
\keyword{internal}
