% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{MKtab<-}
\alias{MKtab<-}
\title{Setzt in \code{character} vector, oder in colnames eines Objekts die Nummer der ZA-BH-Tabelle ein.}
\usage{
MKtab(x) <- value
}
\description{
Setzt in \code{character} vector, oder in colnames eines Objekts die Nummer der ZA-BH-Tabelle ein.
}
\author{
Daniel Hoop
}
