% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{MKtab}
\alias{MKtab}
\title{Extrahiert aus \code{character} vector oder aus colnames eines Objekts die Nummer der ZA-BH-Tabelle. Gibt \code{character} vector zurueck.}
\usage{
MKtab(x)
}
\description{
Extrahiert aus \code{character} vector oder aus colnames eines Objekts die Nummer der ZA-BH-Tabelle. Gibt \code{character} vector zurueck.
}
\author{
Daniel Hoop
}
