% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{month}
\alias{month}
\title{Extract the month from a POSIX time object.}
\usage{
month(time)
}
\arguments{
\item{time}{A POSIX time object.}
}
\value{
The month as numeric.
}
\description{
Extract the month from a POSIX time object.
}
\examples{
month(Sys.time())
}
\author{
Daniel Hoop
}
