% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{nMKcol}
\alias{nMKcol}
\title{Extrahiert aus \code{character} vector oder aus colnames eines Objekts die Nummer der ZA-BH-Spalte. Gibt \code{numeric} vector zurueck.}
\usage{
nMKcol(x)
}
\description{
Extrahiert aus \code{character} vector oder aus colnames eines Objekts die Nummer der ZA-BH-Spalte. Gibt \code{numeric} vector zurueck.
}
\author{
Daniel Hoop
}
