% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{nMKtab}
\alias{nMKtab}
\title{Extrahiert aus \code{character} vector oder aus colnames eines Objekts die Nummer der ZA-BH-Tabelle. Gibt \code{numeric} vector zurueck.}
\usage{
nMKtab(x)
}
\description{
Extrahiert aus \code{character} vector oder aus colnames eines Objekts die Nummer der ZA-BH-Tabelle. Gibt \code{numeric} vector zurueck.
}
\author{
Daniel Hoop
}
