% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{normalize}
\alias{normalize}
\title{Tries to normalize each columng of a given data.frame by applying different functions.}
\usage{
normalize(
  data,
  qq = TRUE,
  window = FALSE,
  funcs = list(`x^(-2)` = function(x) x^(-2), `x^(-1)` = function(x) x^(-1), `x^(-0.5)` =
    function(x) x^(-0.5), `log(x)` = function(x) log(x), `x^(0.5)` = function(x) x^(0.5),
    x = function(x) x, `x^(2)` = function(x) x^(2), `exp(x)` = function(x) exp(x)),
  ...
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} with data to be normalized.}

\item{qq}{Logical value indicating if qq plot should be plotted.}

\item{window}{Logical value indicating if a new window should be opened for the qq plot.}

\item{funcs}{The functions that will be tried in order to make the data "more" normally distributed than it is.}
}
\value{
A list containing
\item{data}{The new data that has potentially been normalized.}
\item{pValsPerCol}{The p values that were calculated by the lillifors test -> to check normal distribution. Values \emph{above} 0.05 indicate normal distribution.}
\item{useFuncsPerCol}{The functions that were applied on all columns of data to come closer to normal distribution.}
}
\description{
Tries to normalize each columng of a given data.frame by applying different functions.
}
\details{
Functions might be applied to different columns even though the lilliefors test did not yield a p.value > 0.05.
This will be the case if the p.value gets larger compared to the original distribution.
}
\author{
Daniel Hoop
}
