% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{paste.cols}
\alias{paste.cols}
\title{Paste all columns of a \code{data.frame} or \code{matrix} and collapse them.}
\usage{
paste.cols(dat, cols = NULL, sep = "_")
}
\arguments{
\item{dat}{The \code{data.frame} or \code{matrix} of which all columns will be collapsed.}

\item{cols}{Optionally: The columns in \code{dat} which should be collapsed.}

\item{sep}{The \code{character} which should be used as a separator between the collapsed column values.}
}
\description{
Paste all columns of a \code{data.frame} or \code{matrix} and collapse them.
}
\author{
Daniel Hoop
}
