% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{perc.change}
\alias{perc.change}
\title{Convert decimal number in "percentage change".}
\usage{
perc.change(x, digits = 1)
}
\description{
Convert decimal number in "percentage change".
}
\author{
Daniel Hoop
}
