% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{pValToSign}
\alias{pValToSign}
\title{Transform p values to sign}
\usage{
pValToSign(pVal, indicate0.1 = FALSE)
}
\arguments{
\item{pVal}{A numeric vector containing p values (from 0 to 1)}

\item{indicate0.1}{Logical value indicating if 0.1 should also be shown as "."}
}
\value{
A character vector indicating significant differences where "***" = 0.001, "**" = 0.01, "*" = 0.05 and "." = 0.1
}
\description{
Transform p values to sign
}
\examples{
pValToSign(c(0.05, 0.049, 0.01, 0.009))
}
