% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{quantile.inverse}
\alias{quantile.inverse}
\title{Inverse quantile function.}
\usage{
\method{quantile}{inverse}(x, value, na.rm = TRUE)
}
\arguments{
\item{x}{All values in the sample.}

\item{value}{The value of which the probability in the cumulative distribution should be calculated.}
}
\description{
The funciton does not yield exactly the same results as you would get using the \code{\link[stats:quantile]{stats::quantile}} function because the distribution functions differ.
}
\author{
Daniel Hoop
}
