% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{quantile.weight.inverse}
\alias{quantile.weight.inverse}
\title{Inverse weighted quantile function.}
\usage{
\method{quantile}{weight.inverse}(x, weights = NULL, value, diff = 0.001, max.iter = 100, na.rm = TRUE)
}
\arguments{
\item{x}{All values in the sample.}

\item{value}{The value of which the probability in the cumulative distribution should be calculated.}

\item{diff}{The maximal relative difference between the real quantile value and the approximated quantile value.}
}
\description{
Finds the approximate quantile probability for a given value.
}
\examples{
probs <- quantile.weight.inverse(x = 1:100, weights = NULL, value = 60, diff = 0.001)
print(probs)
value <- quantile.weight(x = 1:100, weights = NULL, probs = probs)
print(value)
}
\author{
Daniel Hoop
}
