% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{quantile.weight}
\alias{quantile.weight}
\title{Calculate weighted quantiles.}
\usage{
\method{quantile}{weight}(
  x,
  weights = NULL,
  index = NULL,
  probs = 0.5,
  cols = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{x}{The \code{vector}, \code{data.frame}, or \code{matrix} that contains the data for which the quantiles should be calculated.}

\item{weights}{Optionally: A numeric vector containing the weights for each observation given in \code{x}.}

\item{index}{Optionally: A numeric vector or \code{data.frame}/\code{list} containing the indices. See also \code{\link[base:tapply]{base::tapply}}.}

\item{probs}{The probabilities of the quantiles that should be calculated.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{Optionally:}{The columns in \code{x} for which the calculations should be done.}
}
\description{
Calculate weighted quantiles.
}
\author{
Daniel Hoop
}
