% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{rbind.1b1}
\alias{rbind.1b1}
\title{Binds all given matrices in ... row by row.}
\usage{
\method{rbind}{`1b1`}(
  ...,
  add.names = c("char", "num", "obj.names", "own.names", "none"),
  own.names = NULL,
  names.at.front = FALSE,
  sep.sign = "_",
  cbind = FALSE
)
}
\arguments{
\item{add.names}{If add.names="own.names" then own.names must be given.}

\item{sep.sign}{Separates the original rownames from the add.names-endings}

\item{cbind}{Is only used for the \code{\link{cbind.1b1}} function (below) which is some kind of wrapper function for \code{rbind.1b1}.}
}
\description{
Binds all given matrices in ... row by row.
}
\examples{
a <- b <- c <- matrix(1:100,nrow=10,dimnames=list(letters[1:10],letters[11:20]) )
rbind.1b1(a, b, c, add.names="none")
cbind.1b1(a ,b, c, add.names="none")
}
\seealso{
\code{\link{cbind.1b1}}
}
