% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{rdata.to.csv}
\alias{rdata.to.csv}
\title{Write an object that is contained in a RData file into a CSV file that contains the same data.}
\usage{
rdata.to.csv(path)
}
\description{
Write an object that is contained in a RData file into a CSV file that contains the same data.
}
\author{
Daniel Hoop
}
