% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{read.cb}
\alias{read.cb}
\title{Read data from the clipboard.}
\usage{
read.cb(names = c("col", "rowcol", "row", "no"), no.data.frame = FALSE, ...)
}
\arguments{
\item{names}{Dimension names to be saved in the file. \code{"col"} for colnames, \code{"rowcol"} for rownames & colnames, \code{"row"} for rownames, and \code{"no"} for no dimnames.}

\item{...}{Further arguments to be passed to \code{read.table}.}
}
\description{
Read data from the clipboard.
}
\author{
Daniel Hoop
}
