% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{read.xlsx}
\alias{read.xlsx}
\title{Read a sheet from an Excel file}
\usage{
read.xlsx(filename, sheet = 1, ...)
}
\arguments{
\item{filename}{The Excel file.}

\item{sheet}{The name or index of the sheet to read data from.}

\item{...}{Further arguments to be passed into \code{\link[openxlsx:read.xlsx]{openxlsx::read.xlsx}}}
}
\description{
Read a sheet from an Excel file
}
\author{
Daniel Hoop
}
