% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{read.xlsx_XLConnect}
\alias{read.xlsx_XLConnect}
\title{Read a sheet from an Excel file (using the XLConnect package)}
\usage{
read.xlsx_XLConnect(filename, sheet = 1, ...)
}
\arguments{
\item{filename}{
  Filename (absolute or relative) of Excel workbook to be loaded. Supported are Excel '97 (*.xls) and
  OOXML (Excel 2007+, *.xlsx) file formats. Paths are expanded using \code{path.expand}.
}

\item{sheet}{The name or index of the worksheet to read from}
}
\description{
Read a sheet from an Excel file (using the XLConnect package)
}
\author{
Daniel Hoop
}
\keyword{internal}
