% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{readCsvZaData}
\alias{readCsvZaData}
\title{Read ZA data from 'raw' CSV file downloaded from 'Erhebungsbogen'.}
\usage{
readCsvZaData(file, returnRaw = FALSE)
}
\arguments{
\item{file}{The file containing the data as CSV. It looks like this:\cr\cr
\code{betrieb;jahr;merkmal;wert;index;schnittstelle;text;}
\code{1010431;2020;P100_1100_03200;;0;3;Software ABC;}
\code{1010431;2020;P100_1100_03300;;0;3;max.muster@kreditkasse.ch;}
\code{1010431;2020;P100_1100_03400;90;0;3;;}}

\item{returnRaw}{Logical value indicating if the raw table should be returned.}
}
\value{
A data.frame with one row. The column names contain the "Merkmals UIDs".
}
\description{
Read ZA data from 'raw' CSV file downloaded from 'Erhebungsbogen'.
}
\examples{
readCsvZaData("SPA-202000001010431.txt")
}
