% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{readXmlZaData}
\alias{readXmlZaData}
\title{Read ZA data from XML}
\usage{
readXmlZaData(
  file = NULL,
  what = c("farm-mm", "farm-journal", "za-accounts", "za-costunits"),
  transformFarmData = TRUE
)
}
\arguments{
\item{file}{The file containing the data as XML.}

\item{what}{\itemize{
\item \code{"farm-mm"} Return the converted "Merkmale" from the XML file of a single farm.
\item \code{"farm-journal"} Return the journal entries from the XML file of a single farm.
\item \code{"za-accounts"} Return the ZA accounts from the file that is usually named "za-accountsystem.xml"
\item \code{"za-costunits"} Return the ZA cost units from the file that is usually named "za-accountsystem.xml"
}}

\item{transformFarmData}{Logical value. If \code{TRUE}, the farm data will be transformed using \code{\link{transformRawZaMmToDf}}.}
}
\value{
A data.frame
}
\description{
Read ZA data from XML
}
\examples{
readXmlZaData("/home/agsad.admin.ch/f80823148/mnt/Data-Work-RE/25_Agricultural_Economics-RE/251_ZA-BH_protected/ZADaten/SpB/Dateneinlieferung/2020/apr/DCollectZA-2021-40400039905.xml", what = "farm-mm")
readXmlZaData("/home/agsad.admin.ch/f80823148/mnt/Data-Work-RE/25_Agricultural_Economics-RE/251_ZA-BH_protected/ZADaten/SpB/Dateneinlieferung/2020/apr/DCollectZA-2021-40400039905.xml", what = "farm-journal")
readXmlZaData("za-accountsystem.xml", what = "za-accounts") # Path is invalid.
}
