% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{recover.distorted.string}
\alias{recover.distorted.string}
\title{Compares the strings in argument "original" and "distorted".}
\usage{
recover.distorted.string(
  original,
  distorted,
  max.rel.dist = 0.7,
  no.match.NA = TRUE,
  ...
)
}
\arguments{
\item{...}{further arguments to pass into \code{\link[stringist:stringdistmatrix]{stringist::stringdistmatrix}}}
}
\description{
It replaces all occurences in `distorted` with matches in `orgininal` if the relative distance is smaller than `max.rel.dist`.
}
\examples{
recover.distorted.string(
  original = c("Aadorf", "Schüpfen", "Zürich", "Argau", "Lausanne"),
  distorted = c("Adorf", "Sch?pfen", "Bern"), no.match.NA = TRUE)
}
\author{
Daniel Hoop
If `no.match.NA == TRUE`, then no matches will be set to `NA` in the output vector. Otherwise no matches will be the same as given in `distorted`.
}
\keyword{internal}
