% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{rekid.zaid}
\alias{rekid.zaid}
\title{Translate between REK_ID and ZA_ID}
\usage{
rekid.zaid(id, reverse = FALSE, BHJ = NULL, no.match.NA = TRUE)
}
\description{
Translate between REK_ID and ZA_ID
}
\author{
Daniel Hoop
}
