% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{rep.rows.1b1}
\alias{rep.rows.1b1}
\title{Repeat rows of matrix (like \code{\link[base:rep]{base::rep}}), but each row one by one.}
\usage{
\method{rep}{rows.1b1}(x, times)
}
\description{
Repeat rows of matrix (like \code{\link[base:rep]{base::rep}}), but each row one by one.
}
\author{
Daniel Hoop
}
