% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{rep.rows}
\alias{rep.rows}
\title{Repeat rows of matrix (like \code{\link[base:rep]{base::rep}}).}
\usage{
\method{rep}{rows}(x, times)
}
\description{
Repeat rows of matrix (like \code{\link[base:rep]{base::rep}}).
}
\author{
Daniel Hoop
}
