% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{repair.csv}
\alias{repair.csv}
\title{If Excel cannot read a csv file anymore apply this function to the file.}
\usage{
repair.csv(
  filepath,
  header = TRUE,
  remove.middle.rows.cols = FALSE,
  save.file = TRUE,
  print.info = TRUE,
  ...
)
}
\description{
By default only the NA rows (and cols) at the end below (right) of the table are removed.
The function can also be use as a read.table function that automatically removes NA rows and cols.
}
\examples{
repair.csv("filepath", header=TRUE, remove.middle.rows.cols=TRUE, save.file=FALSE)
}
\author{
Daniel Hoop
}
\keyword{internal}
