% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{replace.values}
\alias{replace.values}
\title{Replace values with other values.}
\usage{
replace.values(
  search,
  replace,
  x,
  no.match.NA = FALSE,
  gsub = FALSE,
  fixed = FALSE
)
}
\arguments{
\item{search}{The vector containing values that will be searched.}

\item{replace}{The vector containing values that will be inserted, instead of those, given in `search`.
Each vector place in `search` must correspond to the same vector place in `replace`.}

\item{x}{The vector to do the replacement in.}

\item{no.match.NA}{Logical value indicating if elements in `x` that did not match any element in `search` should either stay as they were in `x` (`no.match.NA = FALSE`),
or should be set to `NA` instead (`no.match.NA = TRUE`).}

\item{gsub}{Logical value indicating if partial strings should be searched, using the function \code{\link[base:gsub]{base::gsub}}. If `FALSE`, then only full string matches will be replaced.}

\item{fixed}{If `gsub = TRUE`, then the characters given in `search` will be interpreted as regular expressions. If you want to avoid this, set `fixed = TRUE`.}
}
\description{
Replace values with other values.
}
\examples{
search <-  c( 1,  2,  9)
replace <- c("a","b","c")
x <- c(NA, 0, 1, 1, 2, 3)
replace.values(search, replace, x)
replace.values(search, replace, x, no.match.NA = TRUE)
replace.values(c("a", "b"), c("1", "2"), c("XX_a_XX", "YY_b_YY"), gsub = TRUE)
}
\seealso{
\code{\link[grep:base]{grep::base}}
}
\author{
Daniel Hoop
}
