% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{require.package}
\alias{require.package}
\title{Install a package if not yet installed. Later require the package.}
\usage{
require.package(..., repos = "http://stat.ethz.ch/CRAN/")
}
\description{
Install a package if not yet installed. Later require the package.
}
\author{
Daniel Hoop
}
