% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{right.to.left}
\alias{right.to.left}
\title{Flips matrix right to left (same as \code{\link{left.to.right}}).}
\usage{
right.to.left(x)
}
\description{
Flips matrix right to left (same as \code{\link{left.to.right}}).
}
\author{
Daniel Hoop
}
\keyword{internal}
