% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{rm.gc.keep}
\alias{rm.gc.keep}
\title{Removes and garbage collects all objects in the current environment (can also be within a function) except for the variables defined in keep (charcter vector).}
\usage{
rm.gc.keep(keep)
}
\arguments{
\item{keep}{A character vector naming all objects that should be kept.}
}
\description{
Removes and garbage collects all objects in the current environment (can also be within a function) except for the variables defined in keep (charcter vector).
}
\author{
Daniel Hoop
}
\keyword{internal}
