% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{round.currency}
\alias{round.currency}
\title{Round to the next 0.05 digit.}
\usage{
\method{round}{currency}(x, stringOut = TRUE)
}
\description{
Round to the next 0.05 digit.
}
\examples{
round.currency(12.31, stringOut = TRUE)
[1] "12.30"
}
\author{
Daniel Hoop
}
\keyword{internal}
