% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{second}
\alias{second}
\title{Extract the second from a POSIX time object.}
\usage{
second(time)
}
\arguments{
\item{time}{A POSIX time object.}
}
\value{
The second as numeric.
}
\description{
Extract the second from a POSIX time object.
}
\examples{
second(Sys.time())
}
\author{
Daniel Hoop
}
