% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{showMissingNames}
\alias{showMissingNames}
\title{Show missing names}
\usage{
showMissingNames(check, against, msgPrefix = NULL)
}
\arguments{
\item{check}{The vector which holds the elements that should be checked.}

\item{against}{The vector or data.frame/matrix which has all available names.}

\item{msgPrefix}{The message prefix to show when an error occurs. After that prefix, all missing names will be listed. If NULL, then the default prefix will be given.}
}
\description{
Show missing names
}
\examples{
cols <- c("a", "b", "c", "d")
data <- data.frame(a = 1, b = 2)
showMissingNames(cols, data)
}
\author{
Daniel Hoop
}
