% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{signif.equally}
\alias{signif.equally}
\title{Round all elements of vector x to the same number of significant digits.}
\usage{
\method{signif}{equally}(x, digits = 3, max = 1, margin = 2)
}
\arguments{
\item{margin}{is only used when a `data.frame`/`matrix` is given. Then \code{\link[base:apply]{base::apply}} is used.}
}
\description{
Round all elements of vector x to the same number of significant digits.
}
\author{
Daniel Hoop
}
\keyword{internal}
