% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{source}
\alias{source}
\title{Like \code{\link[base:source]{base::source}}, but will detect if a file located on Agroscope OS-LW is sourced, and will adapt encoding depending on BIT client / RStudio server.}
\usage{
source(file, encoding = getOption("encoding"), ...)
}
\arguments{
\item{file}{a \link[base]{connection} or a character string giving the
    pathname of the file or URL to read from.  The \code{\link[base]{stdin}()}
    connection reads from the console when interactive.}

\item{encoding}{character vector.  The encoding(s) to be assumed when
    \code{file} is a character string: see \code{\link[base]{file}}.  A
    possible value is \code{"unknown"} when the encoding is guessed: see
    the \sQuote{Encodings} section.}

\item{...}{(for \code{withAutoprint()}:) further (non-file related)
    arguments to be passed to \code{source(.)}.}
}
\description{
Like \code{\link[base:source]{base::source}}, but will detect if a file located on Agroscope OS-LW is sourced, and will adapt encoding depending on BIT client / RStudio server.
}
\seealso{
\code{\link[readr:guess_encoding]{readr::guess_encoding}}
}
\author{
Daniel Hoop
}
