% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{substr.rev}
\alias{substr.rev}
\title{Take a substring but start at the end (reverse substring).}
\usage{
substr.rev(char, start, end, reverse = FALSE)
}
\description{
Take a substring but start at the end (reverse substring).
}
\author{
Daniel Hoop
}
