% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{summary.long}
\alias{summary.long}
\title{'Long' summary, meaning that more quantiles will be calculated, compared to the \code{\link{summary}} function.}
\usage{
\method{summary}{long}(
  x,
  quant = 10,
  digits = 2,
  na.rm = TRUE,
  margin = 2,
  reverse = FALSE,
  ...
)
}
\description{
'Long' summary, meaning that more quantiles will be calculated, compared to the \code{\link{summary}} function.
}
\author{
Daniel Hoop
}
