% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{table.fixed}
\alias{table.fixed}
\title{Puts the result of \code{table(X)} into a fixed given vector with specified names.}
\usage{
table.fixed(..., names.result = NULL, vector.result = FALSE, sep.sign = "_")
}
\description{
It is a wrapper function for \code{\link{tapply.fixed}}.
}
\details{
For information on the arguments see \code{\link{tapply.fixed}}.
}
\examples{
X <- round(runif(100,1,15))
table.fixed(X, names.result=1:5)
table.fixed(round(runif(100,1,15)), round(runif(100,1,15)), names.result=list(1:20, 5:15), vector.result=FALSE)
}
\seealso{
\code{\link{tapply.fixed}}
}
\author{
Daniel Hoop
}
