% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{tapply.fixed}
\alias{tapply.fixed}
\title{Puts the result of tapply(X,INDEX,FUN) into a fixed given vector with names=names.result.}
\usage{
tapply.fixed(
  X,
  INDEX,
  FUN,
  names.result = NULL,
  missing.value = NA,
  vector.result = FALSE,
  sep.sign = "_",
  warn = FALSE
)
}
\arguments{
\item{X}{an \R object for which a \code{\link[base]{split}} method
    exists.  Typically vector-like, allowing subsetting with
    \code{\link[base]{[}}.}

\item{INDEX}{a \code{\link[base]{list}} of one or more \code{\link[base]{factor}}s,
    each of same length as \code{X}.  The elements are coerced to
    factors by \code{\link[base]{as.factor}}.}

\item{FUN}{a function (or name of a function) to be applied, or \code{NULL}.
    In the case of functions like \code{+}, \code{\%*\%}, etc.,
    the function name must be backquoted or quoted.  If \code{FUN} is
    \code{NULL}, tapply returns a vector which can be used to subscript
    the multi-way array \code{tapply} normally produces.}

\item{names.result}{The names of the resulting vector (including all possible 0/NA entries).}

\item{missing.value}{Which value should be put into the resulting vector if there were no entries in \code{INDEX}?}

\item{vector.result}{Should the result be presented in a \code{vector} (one dimension) or in a multidimensional \code{array} (logical)?}

\item{sep.sign}{The sign to separate the new vector index and \code{names.result} if \code{INDEX} is a \code{list}.}

\item{warn}{Logical value indicating if the function should show a warning if some levels in index get lost because they were not given in \code{vector.result}.}
}
\description{
This is especially useful if some entries are missing in INDEX but you want them to be displayed as well!
Otherwise they would be missing in the result of the tapply() function.
}
\examples{
X <- c(1,2,3,4,5,6,7,8,9,10)
INDEX <- LETTERS[c(1,1,1,1,1,2,2,2,2,2)]
tapply.fixed(X, INDEX, mean, names.result = c("A", "B", "C"))
INDEX <- list(LETTERS[c(1,1,1,1,1,2,2,2,2,2)],
              LETTERS[c(2,2,2,2,2,1,1,1,1,1)])
tapply.fixed(X, INDEX, mean, names.result = c("A_A", "A_B", "B_A", "B_B"), vector.result=TRUE)
}
\seealso{
\code{\link[base:tapply]{base::tapply}}
}
\author{
Daniel Hoop
}
