% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transformRawZaMmToDf}
\alias{transformRawZaMmToDf}
\title{Transform raw delivered ZA data to a dataframe.}
\usage{
transformRawZaMmToDf(data)
}
\arguments{
\item{data}{A data.frame containing the columns "UID", "Index" and "Value". It is the data of one farm only.}
}
\value{
The data converted to a data.frame were the colnames are the "Merkmal UIDs" with one row containing the data of the farm.
}
\description{
Transform raw delivered ZA data to a dataframe.
}
\details{
This is a helper function for \code{\link{readXmlZaData}}
}
\examples{
transformRawMmToDf(data.frame(
  UID = c("P220_2020_00100", "P220_2020_00100", "P100_2100_21220"),
  Index = c("1", "2", NA),
  Value = c("300" ,"250", "2")
))
}
