% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.accountName}
\alias{transl.accountName}
\title{Translate the account number of the ZA-BH-Kontenrahmen from number to name.}
\usage{
transl.accountName(x, lang = c("de", "fr"))
}
\arguments{
\item{x}{Numeric vector containing the account numbers.}

\item{lang}{The language, either `"de"` for german, or `"fr"` for french.}
}
\description{
Translate the account number of the ZA-BH-Kontenrahmen from number to name.
}
\author{
Daniel Hoop
}
