% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.EB.MML}
\alias{transl.EB.MML}
\title{Translate Merkmalsbezeichnungen between the Merkmalsliste and the Erhebungsbogen.}
\usage{
transl.EB.MML(x, reverse = FALSE, give.tab = FALSE)
}
\description{
Translate Merkmalsbezeichnungen between the Merkmalsliste and the Erhebungsbogen.
}
\examples{
transl.EB.MML("P430_0100_94000", reverse = FALSE)
}
\author{
Daniel Hoop
}
