% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.kt}
\alias{transl.kt}
\title{Translate the encoding of the ZA-BH Kanton}
\usage{
transl.kt(x, give.tab = FALSE)
}
\description{
Translate the encoding of the ZA-BH Kanton
}
\examples{
transl.kt(c(1, 20, 20, 19))
}
\author{
Daniel Hoop
}
