% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.lbf}
\alias{transl.lbf}
\title{Translate the encoding of the ZA-BH Landbauform.}
\usage{
transl.lbf(x, give.tab = FALSE)
}
\description{
Translate the encoding of the ZA-BH Landbauform.
}
\examples{
transl.lbf(c(4, 2, 2, 3))
}
\author{
Daniel Hoop
}
