% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.mm}
\alias{transl.mm}
\title{Translate Za-BH-Merkmale into names.}
\usage{
transl.mm(x, gsub = FALSE, excel.format = FALSE)
}
\description{
Translate Za-BH-Merkmale into names.
}
\examples{
transl.mm("P430_0100_94000")
}
\author{
Daniel Hoop
}
