% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.ref.210row}
\alias{transl.ref.210row}
\title{Translate the rows in the ZA-BH-Referenzbetriebs-Table 210 to the Betriebszweigs-Namen.}
\usage{
transl.ref.210row(x, give.tab = FALSE)
}
\description{
Translate the rows in the ZA-BH-Referenzbetriebs-Table 210 to the Betriebszweigs-Namen.
}
\examples{
transl.ref.210row("00200")
}
\author{
Daniel Hoop
}
