% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.ref.300row}
\alias{transl.ref.300row}
\title{Translate the rows in the ZA-BH-Referenzbetriebs-Table 300 (Tiere) to the Betriebszweigs-Namen.}
\usage{
transl.ref.300row(x, give.tab = FALSE)
}
\description{
Translate the rows in the ZA-BH-Referenzbetriebs-Table 300 (Tiere) to the Betriebszweigs-Namen.
}
\examples{
transl.ref.300row("00300")
}
\author{
Daniel Hoop
}
