% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.reg}
\alias{transl.reg}
\title{Translate the encoding of the ZA-BH Region.}
\usage{
transl.reg(x, give.tab = FALSE)
}
\description{
Translate the encoding of the ZA-BH Region.
}
\examples{
transl.reg(c(1, 1, 1, 3))
}
\author{
Daniel Hoop
}
