% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.spb.320row}
\alias{transl.spb.320row}
\title{Translate the rows in the ZA-BH-SpB-Table 320 to the Betriebszweigs-Namen.}
\usage{
transl.spb.320row(x, give.tab = FALSE)
}
\description{
Translate the rows in the ZA-BH-SpB-Table 320 to the Betriebszweigs-Namen.
}
\examples{
transl.spb.320row("11110")
}
\author{
Daniel Hoop
}
