% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.spb.330col.340row}
\alias{transl.spb.330col.340row}
\title{Translate bewteen the columns in ZA-BH-SpB-Table 330 and the rows in Table 340.}
\usage{
transl.spb.330col.340row(
  x,
  reverse = FALSE,
  give.tab = FALSE,
  nice.names = FALSE
)
}
\description{
Translate bewteen the columns in ZA-BH-SpB-Table 330 and the rows in Table 340.
}
\examples{
transl.spb.330col.340row(c(2011, 2012))
transl.spb.330col.340row(c(11000), reverse = TRUE)
transl.spb.330col.340row(give.tab = TRUE)
}
\author{
Daniel Hoop
}
