% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.ths}
\alias{transl.ths}
\title{Translate the encoding of the ZA-BH Treuhandstelle.}
\usage{
transl.ths(x, give.tab = FALSE)
}
\description{
Translate the encoding of the ZA-BH Treuhandstelle.
}
\examples{
transl.ths(c(101, 101, 102, 110))
}
\author{
Daniel Hoop
}
