% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.typ}
\alias{transl.typ}
\title{Translate the encoding of the ZA-BH Betriebstyp.}
\usage{
transl.typ(x, short = FALSE, FAT99 = FALSE, give.tab = FALSE)
}
\description{
Translate the encoding of the ZA-BH Betriebstyp.
}
\examples{
transl.typ(c(11, 11, 12))
}
\author{
Daniel Hoop
}
