% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{transl.za}
\alias{transl.za}
\title{Translate ZA-BH encodings}
\usage{
transl.za(
  x,
  name = c("ak_ausbildung", "ak_geschlecht", "ak_rolle", "kanton", "landbauform",
    "region", "sprache", "ths", "typ_s3", "typ_s4", "zone"),
  length = c("long", "short"),
  lang = c("de", "fr", "it", "en"),
  sample = c("unspecific", "spe", "spb", "ref"),
  year = NULL
)
}
\arguments{
\item{x}{The encoded vector that should be decoded.}

\item{name}{The name of the encoding, e.g. \code{"region"}.}

\item{length}{Either the \code{"long"}, or the \code{"short"} version of the decoded value. E.g. "Tal", or "Talregion". Sometimes, the long and the short version do not differ.}

\item{lang}{The language to decode to.}

\item{sample}{The sample for which the decoding should be done.}

\item{year}{Not yet implemented: The year vector which specifies the valid encoding. Each vector place in \code{year} must correspond the same vector place in \code{x}. This could be useful in the future if encodings change.}
}
\description{
Translate ZA-BH encodings
}
\details{
The encodings are managed in the following file: \code{//evdad.admin.ch/AGROSCOPE_OS/2/5/2/1/3/3/4272/MerkmalCodierungen.csv}.
}
\examples{
transl.za(c(1, 2, 3), "region")
transl.za(c(11, 21), "typ_s3", sample = "spe")
transl.za(c(11, 21), "typ_s3", sample = "ref")
}
\author{
Daniel Hoop
}
