% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{trennzeichen1000}
\alias{trennzeichen1000}
\title{FUNKTION IS FEHLERHAFT!!! Fuegt Tausender-Trennzeichen in Zahlen ein.}
\usage{
trennzeichen1000(x, sign = "'", digits = 2, signifequally = FALSE, power = 9)
}
\description{
FUNKTION IS FEHLERHAFT!!! Fuegt Tausender-Trennzeichen in Zahlen ein.
}
\examples{
trennzeichen1000(seq(1.5, 45000000.345, 900000.5))
}
\author{
Daniel Hoop
}
\keyword{internal}
