% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{unregex}
\alias{unregex}
\title{Same as \code{\link{escapeStr}}, but deprecated. Will throw an error.}
\usage{
unregex(x)
}
\description{
Same as \code{\link{escapeStr}}, but deprecated. Will throw an error.
}
\author{
Daniel Hoop
}
