% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{var.weight}
\alias{var.weight}
\title{Calculate weighted variances}
\usage{
var.weight(
  data,
  weights = NULL,
  index = NULL,
  cols = NULL,
  na.rm = TRUE,
  change.output.str = FALSE
)
}
\arguments{
\item{data}{The data of which the weighted variances should be calculated. Can be \code{data.frame}, \code{matrix} or \code{vector}.}

\item{weights}{Optionally: A numeric vector containing the weights for each observation given in \code{x}.}

\item{index}{Optionally: A numeric vector or \code{data.frame}/\code{list} containing the indices. See also \code{\link[base:tapply]{base::tapply}}.}

\item{cols}{Optionally: The columns in data for which the variance should be calculated. If \code{data} contains the columns \code{"a"} and \code{"b"}, then \code{cols} can also be \code{"I(a/b)"}}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{change.output.str}{Logical value indicating if the output structure should be changed in case more than 1 index is given in \code{index}.}
}
\description{
Calculate weighted variances
}
\seealso{
\code{\link{mean.weight}} for examples
}
\author{
Daniel Hoop
}
