% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{vergleichslohn}
\alias{vergleichslohn}
\title{Get Vergleichslohn data (opportunity costs of family labor).}
\usage{
vergleichslohn(region = NULL, jahr = NULL)
}
\arguments{
\item{region}{Optional: The region(s) for which the Vergleichslohn should be returned.}

\item{jahr}{Optional: The year(s) for which the Vergleichslohn should be returned.}
}
\value{
A matrix which differentiates for regions in rows, and for years in columns.
}
\description{
Get Vergleichslohn data (opportunity costs of family labor).
}
\examples{
vergleichslohn()
vergleichslohn(c(1,3), 2019:2020)
}
\author{
Daniel Hoop
}
