% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{vergleichszins}
\alias{vergleichszins}
\title{Get Vergleichszins data ('interest rate for' / 'opportunity cost of' equity).}
\usage{
vergleichszins(jahr = NULL)
}
\arguments{
\item{jahr}{Optional: The year(s) for which the Vergleichszins should be returned.}
}
\value{
A matrix with 1 row containing the interest rate for each year in a separate column.
}
\description{
Get Vergleichszins data ('interest rate for' / 'opportunity cost of' equity).
}
\examples{
vergleichszins()
vergleichszins(2018:2020)
}
\author{
Daniel Hoop
}
