% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{view}
\alias{view}
\title{Creates a CSV file from a data.frame/matrix and opens it with the default CSV-opening-program of the computer.}
\usage{
view(
  x,
  names = c("col", "rowcol", "row", "no"),
  nrows = -1,
  ncols = -1,
  fastViewOfSubset = TRUE,
  folder = NULL,
  filename = NULL,
  quote = FALSE,
  na = "NA",
  sep = "EXCEL",
  decimal.mark = ".",
  openFolder = FALSE,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{data.frame/matrix}

\item{names}{Dimension names to be saved in the file. \code{"col"} for colnames, \code{"rowcol"} for rownames & colnames, \code{"row"} for rownames, and \code{"no"} for no dimnames.}

\item{nrows}{Maximum number of rows to be saved (for higher speed with large datasets). If \code{n = -1}, all rows will be displayed. See also the help for \code{\link[utils:read.table]{utils::read.table}}}

\item{ncols}{Maximum number of columns to be saved (for higher speed with large datasets).}

\item{folder}{Directory, where the temporary file should be saved. If \code{NULL}, then a temporary file will be created using \code{\link[base:tempdir]{base::tempdir}}.}

\item{filename}{Name of the file. If \code{NULL}, then a generic filename will be given.}

\item{quote}{Should quotes be written into the csv File? -> See also the help for \code{\link[utils:write.table]{utils::write.table}}.}

\item{na}{How should NA values be displayed in the csv File? -> See also the help for \code{\link[utils:write.table]{utils::write.table}}.}

\item{sep}{The separator. If \code{"EXCEL"}, then a xlsx file is created and opened.}

\item{openFolder}{Should the folder with all temporary files be opened after having created the file?}

\item{title}{The title to show when \code{\link[utils:View]{utils::View}} is called on the RStudio server.}

\item{...}{Further arguments to be passed to \code{read.table}.}
}
\description{
Creates a CSV file from a data.frame/matrix and opens it with the default CSV-opening-program of the computer.
}
\examples{
x <- matrix(1:10, ncol = 2)
view(x)
}
\author{
Daniel Hoop
}
