% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{viewFarmData}
\alias{viewFarmData}
\title{Showing the data of a farm in a spread sheet like format.}
\usage{
viewFarmData(
  data,
  table,
  filterFunc = function(x) x,
  digits = 2,
  translate = TRUE,
  open = TRUE,
  sep = "EXCEL"
)
}
\arguments{
\item{data}{The data of the farm. Either a data.frame/matrix or a vector. If the data.frame/matrix has more than 1 row, then the first row will be chosen.}

\item{table}{The name of the table, e.g. \code{"P320"}.}

\item{filterFunc}{A function which filters names of figures. E.g. \code{function(x) 10000 <= nMKrow(x) & nMKrow(x) < 50000}}

\item{digits}{The number of digits (for rounding).}

\item{translate}{Logical value indicating if the rows should be translated.}

\item{open}{Logical value indicating if the farm data should be opened (using \code{zaUtils::view}) or only be (invisibly) returned as a matrix.}

\item{sep}{The separation character for the file to be opened (via \code{\link{view}} function). "EXCEL" for xlsx file.}
}
\value{
An invisible matrix.
}
\description{
Showing the data of a farm in a spread sheet like format.
}
\author{
Daniel Hoop
}
