% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{wgs84.to.ch1903}
\alias{wgs84.to.ch1903}
\title{Convert from internation geographic system (WGS84; N, E) to Swiss geographic system (CH1903; y, x)}
\usage{
wgs84.to.ch1903(N, E)
}
\arguments{
\item{N}{A vector c(degree, seconds, minutes) North or a decimal number of length 1 (giving only the degrees).}

\item{E}{A vector c(degree, seconds, minutes) East or a decimal number of length 1 (giving only the degrees).}
}
\value{
The coordinates (x, y) in the Swiss geographic system CH1903.
}
\description{
Convert from internation geographic system (WGS84; N, E) to Swiss geographic system (CH1903; y, x)
}
\examples{
N <- c(46,  2, 38.87)
E <- c( 8, 43, 49.79)
wgs84.to.ch1903(N, E)
c(x = 1e+05, y = 7e+05)
}
