% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{wide.to.long.df}
\alias{wide.to.long.df}
\title{Erstellt aus einem Breiten data.frame, wo die Betriebszweige in verschiedenen Spalten stehen einen langen wo alle rawIndex untereinander stehen und indexiert sind.}
\usage{
wide.to.long.df(
  fixed = NULL,
  transform,
  rawIndex = colnames(transform),
  colnameIndex = "index",
  colnameValue = "value"
)
}
\arguments{
\item{fixed}{Datensatz welcher die Betriebsinformationen enthaelt (z.B. `"ID"` und `"Jahr"`). Kann auch `NULL` sein. Dann werden keine Zusatzinfos angehaengt.}

\item{transform}{= Datensatz, in welchem alle `rawIndex` nebeneinander in Spalten stehen}

\item{rawIndex}{Index-Rohling}

\item{colnameIndex}{Name der Index-Spalte im neuen Datensatz.}

\item{colnameValue}{Name der neuen Spalte, in die die Werte reingeschrieben werden.}
}
\description{
Das ist vor allem noetig fuer die Konversion von Datensatzen im Skript 020_aggreg_ME_param.R
}
\seealso{
\code{\link[stats:reshape]{stats::reshape}}

\code{\link[reshape:melt]{reshape::melt}}, \code{\link[stats:reshape]{stats::reshape}}, \code{\link{long.to.wide.df}}
}
\author{
Daniel Hoop
}
