% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{wilcox.or.chisq}
\alias{wilcox.or.chisq}
\title{Calculate the Wilcoxon rank sum statistics for continuous variables and the chi-squared statistics for binary variables (automatically).}
\usage{
wilcox.or.chisq(y, trt, digits = 3, pValuesOnly = TRUE, verbose = TRUE)
}
\arguments{
\item{y}{A numeric vector or data.frame}

\item{trt}{The binary treatment vector}

\item{digits}{The number of digits to round the p.value for the formatted output.}

\item{pValuesOnly}{Default: TRUE. Logical value indicating if only the p value or also signs like \code{".", "*", "**", "***"} should be returned as result.}

\item{verbose}{Default: TRUE. Logical value indicating if the detection of binary variables (and the application of the chisq test) should be commented during calculations.}
}
\value{
A numeric vector if \code{pValuesOnly == TRUE}.
Otherwise a list containing the p.values "p.value", signs "sign", and both of them put together "formatted".
}
\description{
Calculate the Wilcoxon rank sum statistics for continuous variables and the chi-squared statistics for binary variables (automatically).
}
\examples{
wilcox.or.chisq(
  data.frame(nonSignVar = rnorm(1:100), signVar = c(rep(1,50),rep(2,50))),
  trt = c(rep(1,50), rep(2,50)))
}
\author{
Daniel Hoop
}
