% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{withTimeout}
\alias{withTimeout}
\title{Evaluate an R expression and interrupts it if it takes too long}
\usage{
withTimeout(
  expr,
  substitute = TRUE,
  envir = parent.frame(),
  timeout,
  cpu = timeout,
  elapsed = timeout,
  onTimeout = c("silent", "error", "warning")
)
}
\arguments{
\item{expr}{The R expression to be evaluated.}

\item{substitute}{If TRUE, argument expr is substitute():ed, otherwise not.}

\item{envir}{The environment in which the expression should be evaluated.}

\item{timeout}{A numeric specifying the maximum number of seconds the expression is allowed to run before being interrupted by the timeout. The cpu and elapsed arguments can be used to specify whether time should be measured in CPU time or in wall time.}

\item{cpu}{A numeric specifying the maximum number of seconds the expression is allowed to run before being interrupted by the timeout. The cpu and elapsed arguments can be used to specify whether time should be measured in CPU time or in wall time.}

\item{elapsed}{A numeric specifying the maximum number of seconds the expression is allowed to run before being interrupted by the timeout. The cpu and elapsed arguments can be used to specify whether time should be measured in CPU time or in wall time.}

\item{onTimeout}{A character specifying what action to take if a timeout event occurs.}
}
\description{
Evaluate an R expression and interrupts it if it takes too long
}
\details{
This function is a slightly adapted version of R.utils::withTimeout (https://cran.r-project.org/package=R.utils)
}
\seealso{
https://cran.r-project.org/package=R.utils
}
