% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{write.xlsx}
\alias{write.xlsx}
\title{Writes an Excel file from either a matrix/data.frame or from a list conaining several matrices/data.frames.}
\usage{
write.xlsx(
  data,
  file,
  sheetName = NULL,
  row.names = FALSE,
  col.names = TRUE,
  append = FALSE,
  asFormula = FALSE,
  wrapText = FALSE,
  columnWidth = -1,
  convertColumnWith = c("default", "pixel", "no")
)
}
\arguments{
\item{data}{A matrix/data.frame or list containing matrices/data.frames}

\item{file}{The path to the xlsx file}

\item{sheetName}{Optional sheet names in the Excel file. If \code{NULL}, then names Sheet1, Sheet2, etc. are given.}

\item{row.names}{Logical value indicating if rownames should be written or character giving the header for rowname colum. Single value or list.}

\item{col.names}{Logical value indicating if a header should be written. Single value or list.}

\item{append}{Logical value indicating if the excel file should be appended rather than newly created.}

\item{asFormula}{Logical value indicating if formulas should be written rather than values. Note that no \code{"="} sign at the beginning of the cell is required.}

\item{wrapText}{Logical value indicating if the text in the cells should be wrapped (Zeilenumbruch).}

\item{columnWidth}{The column width that will be applied to all columns. If -1, then auto-width is applied.}

\item{convertColumnWith}{If \code{"no"}, then the standard of XLConnect is used. It is the 1/256 part of a character. \code{"default"} will convert in 1 character units. \code{"pixel"} will convert to number of pixel.}
}
\description{
It is a convenient version of writeWorksheetToFile{XLConnect}
}
\author{
Daniel Hoop
}
