% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{xml2NodesToMatrix}
\alias{xml2NodesToMatrix}
\title{Convert xml2 nodes to a matrix.}
\usage{
xml2NodesToMatrix(nodes)
}
\arguments{
\item{nodes}{The xml nodes.}
}
\value{
A data.frame
}
\description{
Convert xml2 nodes to a matrix.
}
\examples{
xml <- "
<Root>
  <Attributes>
    <Attribute><name>a</name><value>1</value></Attribute>
    <Attribute><name>b</name><value>2</value></Attribute>
  </Attributes>
</Root>"

xml2NodesToMatrix(xml2::xml_find_all(xml2::read_xml(xml), ".//Attribute"))
}
