% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{year}
\alias{year}
\title{Extract the year from a POSIX time object.}
\usage{
year(time)
}
\arguments{
\item{time}{A POSIX time object.}
}
\value{
The year as numeric.
}
\description{
Extract the year from a POSIX time object.
}
\examples{
year(Sys.time())
}
\author{
Daniel Hoop
}
