% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaUtils.R
\name{plm.within.between}
\alias{plm.within.between}
\title{Wrapper for \code{wibe::wibe}.}
\usage{
plm.within.between(
  data,
  index = NULL,
  additionalIndex = NULL,
  randomEffects = NULL,
  Y,
  timeVariantX,
  timeInvariantX = NULL,
  timeDummies = NULL,
  weights = NULL,
  dfEstimator = c("Satterthwaite", "Kenward-Roger"),
  vcovFunc = NULL,
  usePlmPackageOnly = FALSE,
  useFrmpdPackage = FALSE,
  ...
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing all relevant variables.}

\item{index}{Colnames of the index columns in \code{data}. \code{length(index) == 2} is required. First: individual, second: time. E.g. \code{c("id","year")}. Can be \code{NULL}. But then \code{randomEffects} must be specified.}

\item{randomEffects}{The colnames of the random effects. Can be used to model multiple (nested effects). Nested effects are detected automatically by the lmer function, so just provide a charcter vector containing the column names. Can be \code{NULL}. But then \code{index} must be specified.}

\item{Y}{Colname of the dependent (y) variable.}

\item{timeVariantX}{Colname(s) of the indepentent variables that vary over time within each observation.}

\item{timeInvariantX}{Optional: Colname(s) of the indepentent variables that don't vary over time within each observation. Factors will be converted to dummies.}

\item{timeDummies}{Optional: Colname(s) of the time dummies. Can also be the colname of a factor variable, which will be converted internally to dummies.}

\item{weights}{The weights for each obervation. If \code{NULL}, then it is calculated as the inverse occurence rate of each observation in each random effect. If this should be prevented, then specify \code{weights = rep(1,nrow(data))}.}

\item{dfEstimator}{The method to estimate the degrees of freedom. Either \code{"Satterthwaite"} or \code{"Kenward-Roger"}. See \code{\link[lmerTest:summary.lmerModLmerTest]{lmerTest:::summary.lmerModLmerTest}} for more information.}

\item{vcovFunc}{*Experimental*: A function to calculate the variance covariance matrix. Default is \code{NULL}, and it does not need to be specified because the matrix will be calculated anyway. However, to deal with issues such as heteroskedasticity, a special function such as \code{\link[sandwich:vcovHC]{sandwich::vcovHC}} can be provided.
If additional arguments must be specified in the variance covariance function, then it has to be done along the lines of: \code{vcovFunc = function (x) sandwich::vcovHC(x, type = "HC3")}.}

\item{usePlmPackageOnly}{Logical value indicating if only the plm package should be used, instead of packages like lmerTest and MuMIn.}

\item{useFrmpdPackage}{Logical value indicating if the function \code{\link[frmpd:frmpd]{frmpd::frmpd}} should be used. This package is needed it the dependent (Y) variable is a quotient in the unit interval (i.e. between 0 and 1).}

\item{...}{Further arguments to be passed into \code{\link[plm:plm]{plm::plm}}, \code{\link[lmerTest:lmer]{lmerTest::lmer}}, or \code{\link[frmpd:frmpd]{frmpd::frmpd}}.}
}
\description{
Consider switching package, i.e. working with \code{wibe::wibe} directly.
}
